OAlg.Control.HNFData

Description

reducing a value to head normal form. This is much weaker then Control.DeepSeq.

Documentation

data HNFValue x

values in head normal form.

Constructors

HNFValue x

forall e.Exception e => Failure e

fromHNFValue :: HNFValue x -> x

from head normal form.

Property Let x' be in HNFValue x then holds:

If x' matches HNFValue x then the result of fromHNFValue x' is x.

If x' matches Failure e then evaluation fromHNFValue x' will end up in a throwing e.

class HNFData x where

data reducible to head normal form.

Methods

rhnf :: x -> ()

tries to reduce a value to its head normal form, throwing an Exception for undefined values.

hnfValue :: HNFData x => x -> IO (HNFValue x)

tries to reduce a value x to its head normal form. If the reduction ends up in a SomeException e then this will be catched and Failure e will be returned, otherwise @HNFValue will be returned.