OAlg.Control.Solver

Description

A monad to express solving algebraic problems which may fail.

Note To express algebraic unsolvable - e.g determine the multiplicative inverse of 0 - one should only use the failure mechanism from this module (see examples in solvable).

Result Type For Solving Algebraic Problems

data Solver x

monad to solve algebraic problems which may fail because it is not solvable or ...

failure :: Exception e => e -> Solver x

throwing exception in to Solver to express unsolvable.

handle :: Exception e => Solver x -> (e -> Solver x) -> Solver x

handling exception.

Note Only exceptions expressed via the failure mechanism can be handled by handle!

solve :: Solver x -> x

extracting the solution from the solver. If during solving an exception e occurs, then e will be thrown via the regular throw mechanism of Exception.

Note solve extracts the solution lazy!

solvable :: Solver r -> Bool

checks for solvability of a algebraic problem represented be a solver s (see also solve).