OAlg.Control.Verbose

Description

verbosity on showing.

Verbose

class Show a => Verbose a where

shows a in the context of verbosity.

Minimal complete definition

Nothing

Methods

vshow :: Verbosity -> a -> String

the default implementation is: vshow v a = vshowStr (mnString v) (show a)

data Verbosity

kinds of verbosity.

Constructors

Low

Middle

High

Full

Pretty

vshowStr :: Maybe Int -> String -> String

verbosely showing a string by the given length.

Example

>>> vshowStr (Just 3) "123456789"
"123.."

>>> vshowStr Nothing "123456789"
"123456789"

mnString :: Verbosity -> Maybe Int

default length for a string representation in context of verbosity.

vshowList :: Verbose a => Verbosity -> Maybe Int -> String -> String -> [a] -> String

verbosely showing a list by the given length.

Examples

>>> vshowList Full (Just 3) "[" "]" "abcdef"
"['a','b','c'..]"

>>> vshowList Low (Just 3) "{" "}" ["abcdef","ghijklmn","op","qrst","uvwxyz"]
"{['a','b'..],['g','h'..],['o','p']..}"

mnList :: Verbosity -> Maybe Int

default number of entries for a list representation in context of verbosity.

Percent

newtype Percent x

tagging for showing percentage of a Double.

Constructors

Percent x

showPercent :: Int -> Double -> String

showing a double as percent with the given precision.

Example

>>> showPercent 2 0.912837
" 91.28%"