OAlg.Data.Boolean.Definition

Description

boolean structure for multivalent logic.

Boolean

class Boolean b where

types with a Boolean structure, allowing multivalent logic.

Note Every Enum type which is also Bounded has a natural implementation as false = minBound, true = maxBound, (||) = max, (&&) = min (as there are min and max bounds the operator (||) and (&&) should be implemented with a lazy variant of min and max) and not b = toEnum (fromEnum maxBound - fromEnum t).

Minimal complete definition

true, false, not, ((||), (&&) | or, and)

Methods

false :: b

true :: b

not :: b -> b

(||) :: b -> b -> b infixr 2

or :: [b] -> b

(&&) :: b -> b -> b infixr 3

and :: [b] -> b

(~>) :: b -> b -> b infixr 1

(<~>) :: b -> b -> b infixr 1

eqvl :: [b] -> b

Bool

data Bool

Constructors

False

True

otherwise :: Bool

Structure

data Bol

type representing Boolean structures.