OAlg.Data.Canonical

Description

canonical mappings between two types.

Canonical Mappings

class Embeddable a b where

canonical embedding from a in to b.

Property

inj is injective.

if the two types a and b are also Projectible a b then prj . inj is the identical mapping.

Methods

inj :: a -> b

canonical injetion from a in to b.

class Projectible a b where

canonical projection from b on to a.

Property

prj is surjective.

if the two types a and b are also Projectible a b then prj . inj is the identical mapping.

Methods

prj :: b -> a

canonical projection from b on to a