OAlg.Data.Opposite

Description

predicate for the opposite.

Op

newtype Op x

Predicate for the opposite of a type x.

Constructors

Op x

fromOp :: Op x -> x

from Op x.

fromOpOp :: Op (Op x) -> x

from Op (Op x).

Op2

newtype Op2 h x y

Predicat for the opposite of a two parametrized type h where the two parameters x and y are switched

Constructors

Op2 (h y x)