OAlg.Data.Reducible

Description

reducing values to there canonical value.

Reducible

class Reducible e where

types admitting reducible values.

Definition reduce e is called the algebraic value of e.

Reducing an e twice yield the 'same' value and the idea is that in an algebraic calculation it will be 'safe' to substitute any occurrence of e by its reduced value, i.e. both calculations will yield the same result.

Property Let e be a reducible type admitting equality, then for all e in e holds: reduce (reduce e) == reduce e.

Minimal complete definition

Nothing

Methods

reduce :: e -> e

reducing e to its algebraic value.

Note The default implementation is reduce = id.

reduceWith :: (x -> Rdc x) -> x -> x

reduces x by the given rules until no more reductions are applicable.

(>>>=) :: (x -> Rdc x) -> (x -> Rdc x) -> x -> Rdc x infixr 1

composition of two reductions.

type Rdc = Action RdcState

Action according to the state type RdcState.

data RdcState

reduction state.

Constructors

Unchanged
no reduction has been applied.

Changed
a reduction has been applied.

reducesTo :: x -> Rdc x

indicates that a term has the given reduction step, i.e. returns the given value and sets the state to Changed.