OAlg.Data.Show

Description

showing data with some auxiliary functions.

Show

class Show a where

Minimal complete definition

showsPrec | show

Methods

showsPrec :: Int -> a -> ShowS

show :: a -> String

showList :: [a] -> ShowS

tween :: a -> [a] -> [a]

inserting the given value in between the elements of the given list.

Examples

>>> tween ',' "12345"
"1,2,3,4,5"

>>> tween ',' ""
""

>>> tween ',' "1"
"1"

jtween :: [a] -> [[a]] -> [a]

inserting the given list in between the elements of the given list and joining the result.

Example

>>> jtween ";" ["abcd","efg"]
"abcd;efg"

class Show1 p where

showable for one parameterized types.

Minimal complete definition

Nothing

Methods

show1 :: p x -> String

default show1 :: Show (p x) => p x -> String

class Show2 h where

showable for two parameterized types.

Note We use this class mearly in the context of Path.

Minimal complete definition

Nothing

Methods

show2 :: h a b -> String

default show2 :: Show (h a b) => h a b -> String

type String = [Char]

data Char

Read

class Read a

Minimal complete definition

readsPrec | readPrec

read :: Read a => String -> a