OAlg.Data.Singleton

Description

singleton types having exactly one value.

Singleton

class Singleton s where

types s with exactly one value which is called the unit of s.

Methods

unit :: s

class Singleton1 s where

one parameterized types s with exactly one element for each x which is called the unit1 of s x.

Methods

unit1 :: s x