OAlg.Data.Singular

Description

parameterized types with at most on constructor.

Documentation

class Eq1 p => Singular p

parameterized types with at most on constructor.

Definition A type function p is called singular if and only if for each type x there is at most one constructor for p x.

(<==>) :: Singular p => p x -> p x -> Bool

equivalence of witnesses in p x.

Note p <==> q returns either True or an ImplementationError will be thrown.