OAlg.Data.Statistics

Description

a tool kit for making statistics.

Statistic

mkStatisticW :: Ord x => [x -> String] -> [(Int, x)] -> (Int, [(Double, [String], x)])

makes the statistics of the given list of wight, value pairs.

mkStatistic :: Ord x => [x -> String] -> [x] -> (Int, [(Double, [String], x)])

makes the statistics of the given list of values.

putStatisticW :: (Show x, Ord x) => [x -> String] -> [(Int, x)] -> IO ()

puts the statistics of the given list of wight, values pairs.

putStatistic :: (Show x, Ord x) => [x -> String] -> [x] -> IO ()

puts the statistics of the given list of values.