OAlg.Data.Validable

Description

validable values x, which can be validated via validate (valid x).

Validable

class Validable a where

validation of a value of a.

Methods

valid :: a -> Statement

rnfValid :: NFData x => x -> Statement

validation of being reducible to normal form.

class Validable1 p where

validation of a value of p x.

Minimal complete definition

Nothing

Methods

valid1 :: p x -> Statement

default valid1 :: Validable (p x) => p x -> Statement

class Validable2 h where

validation of a value of h x y.

Minimal complete definition

Nothing

Methods

valid2 :: h x y -> Statement

default valid2 :: Validable (h x y) => h x y -> Statement

XStandard

class Validable x => XStandard x where

standard random variable for x.

Property For all x in the range of xStandard holds: valid x.

Methods

xStandard :: X x

relXStandard :: XStandard x => p x -> Statement

validity of the standard random variable associated to x (p x just serves as proxy and will not be evaluated).