OAlg.Data.X

Description

Random variables for stochastical validation.

Random Variable

data X x

random variable over x, possibly XEmpty. Let x be a type and xx in X x, then we use the idiom x is in the range of xx if there exist a o in Omega such that x is an element of samples xx o.

Note

For the empty set O there is exactly one sigma algebra, i.e. the power set of the empty set O, and for every set X there is exactly one measurable function O -> X, i.e. the empty function, and hence exactly one random variable over O.

To not run into non terminating programs, we restrict the implementation of xa >>= f to a maximal number of iterations to find a suitable sample in xa for which f a is not empty. If the iterations exceed this maximum number, a ProbablyEmpty exception will be thrown.

Constructors

XEmpty

samples :: X x -> Omega -> [x]

infinite list of randomly picked samples of xx according to a initial omega o. If xx is empty then the result will be '[]'.

getSamples

Arguments

:: N
length of the returned list

-> X x

-> IO [x]

gets a list of randomly picked samples.

sample :: X x -> Omega -> x

the first element of samples xx o. If xx is empty then a IsEmpty exception will be thrown.

Statistics

meanValue :: Fractional x => Int -> X x -> Omega -> x

the mean value of n-samples according the state s.

Omega

data Omega

A possible state of the world. It is used for run or samples to generate randomly values.

mkOmega :: Int -> Omega

makes a state.

getOmega :: IO Omega

gets randomly a state.

X

xOmega :: X Omega

random variable of Omega.

xInt :: X Int

uniformly distributed random variable of Ints.

xIntB :: Int -> Int -> X Int

uniformly distributed random variable of Ints in the given range. If the lower bound is greater then the upper bound the result will be XEmpty.

xWord :: X Word

uniformly distributed random variable of Words.

xWordB :: Word -> Word -> X Word

uniformly distributed random variable of Words in the given range. If the lower bound is greater then the upper bound the result will be XEmpty.

xInteger :: X Integer

uniformly distributed random variable of Integers.

xIntegerB :: Integer -> Integer -> X Integer

uniformly distributed random variable of Integers in the given range. If the lower bound is greater then the upper bound the result will be XEmpty.

xChar :: X Char

uniformly distributed random variable of Chars.

xCharB :: Char -> Char -> X Char

uniformly distributed random variable of Chars in the given range. If the lower bound is greater then the upper bound the result will be XEmpty.

xDouble :: X Double

uniformly distributed random variable of Doubles.

xDoubleB :: Double -> Double -> X Double

uniformly distributed random variable of Doubles in the given range. If the lower bound is greater then the upper bound the result will be XEmpty.

xEnum :: (Enum a, Bounded a) => X a

uniformly distributed random variable of a Bounded Enum in the range minBound to maxBound.

xEnumB :: Enum a => a -> a -> X a

uniformly distributed random variable of a Enum in the given range. If the lower bound is greater then the upper bound the result will be XEmpty.

xBool :: X Bool

uniformly distributed random variable of Bools.

xTupple2 :: X a -> X b -> X (a, b)

random variable for pairs.

xTupple3 :: X a -> X b -> X c -> X (a, b, c)

random variable for triples.

xTakeN :: N -> X x -> X [x]

random variable of list with the given length for non empty random variables. Otherwise the result will be XEmpty.

xTakeB :: N -> N -> X x -> X [x]

random variable of lists with a length between the given bounds.

xList :: [X x] -> X [x]

random variable of list.

xOneOf :: [a] -> X a

xOneOf xs is the random variable of xs in xs with a uniformly distribution of the xis, where 0 < length xs. If xs == [] then XEmpty will be the result.

xOneOfX :: [X a] -> X a

as xOneOf.

xOneOfW :: [(Q, a)] -> X a

xOneOfW [(w1,x1)..(wn,xn)] is the random variable of xs in [x1,x2,..xn] with a distribution of the xis of pi = wi/s, where 0 < n, s = w1+w2+..+wn and 0 <= wi for i = 1..n. If n == 0 then XEmpty will be the result.

xOneOfXW :: [(Q, X a)] -> X a

as xOneOfW.

xN :: X N

uniformly distributed random variable in the given range.

xNB :: N -> N -> X N

uniformly distributed random variable in the given range. If the lower bound is greater then the upper bound the result will be XEmpty.

xZ :: X Z

uniformly distributed random variable of Z.

xZB :: Z -> Z -> X Z

uniformly distributed random variable of Z in the given bounds. If the lower bound is greater then the upper bound the result will be XEmpty.

xQ :: X Q

uniformly distributed random variable of Q.

Tools

sum' :: Num x => [x] -> x

a strict and head recursive version of sum.

putDistribution :: (Show x, Ord x) => Int -> X x -> Omega -> IO ()

puts the distribution according of the given number of samples.

putDistribution' :: (Show x, Ord x) => [x -> String] -> Int -> X x -> Omega -> IO ()

puts the distribution according to the given aspects and the given number of samples.

putDistributionIO :: (Show x, Ord x) => Int -> X (IO x) -> Omega -> IO ()

puts the distribution of according the given number of samples.

putDstr :: (x -> [String]) -> Int -> X x -> IO ()

puts the distribution according of the given number of samples.

aspCnstr :: Show x => x -> String

showing the constructor as an aspect.

Exception

data XException

Exceptions for random variables.

Constructors

ProbablyEmpty String

IsEmpty