OAlg.Entity.Diagram.Definition

Description

definition of Diagrams on Oriented structures.

Diagram

data Diagram t n m a where

diagram for a Oriented structure a of type t having n points and m arrows.

Properties Let d be in Diagram t n m a for a Oriented structure a, then holds:

If d matches DiagramChainTo e as then holds: e == end a0 and start ai == end ai+1 for all i = 0..m-2 where a0:|..:|ai:|..:|am-1:|Nil = as.

If d matches DiagramChainFrom s as then holds: s == start a0 and end ai == start ai+1 for all i = 0..m-2 where a0:|..:|ai:|..:|am-1:|Nil = as.

If d matches DiagramParallelLR l r as then holds: orientation a == l:>r for all a in as.

If d matches DiagramParallelRL l r as then holds: orientation a == r:>l for all a in as.

If d matches DiagramSink e as then holds: e == end a for all a in as.

If d matches DiagramSource s as then holds: s == start a for all a in as.

If d matches DiagramGeneral ps aijs then holds pi == start aij and pj == end aij for all aij in aijs and ps = p0..pn-1@.

Constructors

DiagramEmpty :: Diagram 'Empty N0 N0 a

DiagramDiscrete :: FinList n (Point a) -> Diagram Discrete n N0 a

DiagramChainTo :: Point a -> FinList m a -> Diagram (Chain To) (m + 1) m a

DiagramChainFrom :: Point a -> FinList m a -> Diagram (Chain From) (m + 1) m a

DiagramParallelLR :: Point a -> Point a -> FinList m a -> Diagram (Parallel LeftToRight) N2 m a

DiagramParallelRL :: Point a -> Point a -> FinList m a -> Diagram (Parallel RightToLeft) N2 m a

DiagramSink :: Point a -> FinList m a -> Diagram (Star To) (m + 1) m a

DiagramSource :: Point a -> FinList m a -> Diagram (Star From) (m + 1) m a

DiagramGeneral :: FinList n (Point a) -> FinList m (a, Orientation N) -> Diagram General n m a

data DiagramType

the types of a Diagram.

Constructors

Empty

Discrete

Chain Site

Parallel Direction

Star Site

General

rt' :: forall (t :: DiagramType). (Dual (Dual t) :~: t) -> Dual (Dual (Dual t)) :~: Dual t

Dual is well defined on diagram types.

dgType :: Diagram t n m a -> DiagramType

the type of a diagram.

dgTypeRefl :: Diagram t n m a -> Dual (Dual t) :~: t

reflexivity of the underlying diagram type.

dgPoints :: Oriented a => Diagram t n m a -> FinList n (Point a)

the points of a diagram.

dgCenter :: Diagram (Star t) n m c -> Point c

the center point of a Star-diagram.

dgArrows :: Diagram t n m a -> FinList m a

the arrows of a diagram.

dgMap :: Hom Ort h => h a b -> Diagram t n m a -> Diagram t n m b

mapping of a diagram via a homomorphism on Oriented structures.

dgQuiver :: Diagram t n m a -> Quiver n m

the underlying quiver of a diagram.

Chain

chnToStart :: Oriented a => Diagram (Chain To) n m a -> Point a

the last point of the chain.

chnFromStart :: Diagram (Chain From) n m a -> Point a

the first point of the chain.

Parallel

dgPrlAdjZero :: Distributive a => Diagram (Parallel LeftToRight) n m a -> Diagram (Parallel LeftToRight) n (m + 1) a

adjoins a zero arrow as the first parallel arrow.

dgPrlTail :: Diagram (Parallel d) n (m + 1) a -> Diagram (Parallel d) n m a

the _tail__ of a parallel diagram.

dgPrlDiffHead :: Abelian a => Diagram (Parallel d) n (m + 1) a -> Diagram (Parallel d) n (m + 1) a

subtracts to every arrow of the parallel diagram the first arrow.

dgPrlDiffTail :: Abelian a => Diagram (Parallel d) n (m + 1) a -> Diagram (Parallel d) n m a

subtracts the first arrow to all the others an drops it.

Duality

dgToOp :: DiagramDuality f g a -> f a -> g (Op a)

to g (Op a).

dgFromOp :: DiagramDuality f g a -> g (Op a) -> f a

from g (Op a).

data DiagramDuality f g a where

Op-duality between diagrams.

Constructors

DiagramDuality :: Oriented a => (f a :~: Diagram t n m a) -> (g (Op a) :~: Dual (Diagram t n m a)) -> (Dual (Dual t) :~: t) -> DiagramDuality f g a

coDiagram :: Diagram t n m a -> Dual (Diagram t n m a)

the co diagram with its inverse coDiagramInv.

Property Let d be in Diagram t n m a for a Oriented structure a then holds: dgPoints (coDiagram d) == dgPoints d.

coDiagramInv :: Oriented a => (Dual (Dual t) :~: t) -> Dual (Diagram t n m a) -> Diagram t n m a

from the dual diagram, with inverse of coDiagram.

dgFromOpOp :: Oriented a => Diagram t n m (Op (Op a)) -> Diagram t n m a

from Op . Op.

SomeDiagram

data SomeDiagram a where

some diagram.

Constructors

SomeDiagram :: Diagram t n m a -> SomeDiagram a

sdgMap :: Hom Ort h => h a b -> SomeDiagram a -> SomeDiagram b

mapping of some diagram via a homomorphism on Oriented structures.

sdgFromOpOp :: Oriented a => SomeDiagram (Op (Op a)) -> SomeDiagram a

from Op . Op.

coSomeDiagram :: SomeDiagram a -> Dual (SomeDiagram a)

the dual of some diagram, with inverse coSomeDiagramInv.

coSomeDiagramInv :: Oriented a => Dual (SomeDiagram a) -> SomeDiagram a

from the dual of some diagram, with inverse coSomeDiagram.

X

data XDiagram t n m a where

generator for random variables of diagrams.

Constructors

XDiagramEmpty :: XDiagram 'Empty N0 N0 a

XDiagramDiscrete :: Any n -> X (Point a) -> XDiagram Discrete n N0 a

XDiagramChainTo :: Any m -> XOrtSite To a -> XDiagram (Chain To) (m + 1) m a

XDiagramChainFrom :: Any m -> XOrtSite From a -> XDiagram (Chain From) (m + 1) m a

XDiagramParallelLR :: Any m -> XOrtOrientation a -> XDiagram (Parallel LeftToRight) N2 m a

XDiagramParallelRL :: Any m -> XOrtOrientation a -> XDiagram (Parallel RightToLeft) N2 m a

XDiagramSink :: Any m -> XOrtSite To a -> XDiagram (Star To) (m + 1) m a

XDiagramSource :: Any m -> XOrtSite From a -> XDiagram (Star From) (m + 1) m a

xDiagram :: Oriented a => (Dual (Dual t) :~: t) -> XDiagram t n m a -> X (Diagram t n m a)

the induced random variables of diagrams.

xSomeDiagram :: Oriented a => X SomeNatural -> XOrtSite To a -> XOrtSite From a -> XOrtOrientation a -> X (SomeDiagram a)

the induced random variable of some diagrams.

dstSomeDiagram :: Oriented a => Int -> X (SomeDiagram a) -> IO ()

distribution of a random variable of some diagrams.

xSomeDiagramOrnt :: Entity p => X SomeNatural -> X p -> X (SomeDiagram (Orientation p))

random variable of some diagram of Orientation p.