OAlg.Entity.Matrix.Entries

Description

entries of matrices and viewing them as a column of rows respectively as a row of columns.

Entries

newtype Entries i j x

two dimensional partial sequence.

Constructors

Entries (PSequence (i, j) x)

etsxs :: Entries i j x -> [(x, i, j)]

underlying list of indexed entries.

etsEmpty :: Entries i j x

the empty sequence of entries.

etsAdd :: (Additive x, Ord i, Ord j) => Entries i j x -> Entries i j x -> Entries i j x

adding two entries.

Property Let zs = etsAdd xs ys, then holds:

Pre

For all (i,j) in (i,j) where there exists an (x,i,j) in xs and a (y,i,j) in ys holds: root x == root y.

Post

zs is valid.

For all (i,j) in (i,j) holds:

If exists a (x,i,j) in xs but not exists a (y,i,j) in ys then there exists a (z,i,j) in zs with z == x.

If exists a (y,i,j) in ys but not exists a (x,i,j) in xs then there exists a (z,i,j) in zs with z == y.

If exists a (x,i,j) in xs and (y,i,j) in ys then there exists a (z,i,j) in zs with z == x + y.

etsMlt :: (Distributive x, Ord k) => Col i (Row k x) -> Row j (Col k x) -> Col i (Row j x)

multiplication.

etsJoin :: (i ~ N, j ~ N) => ProductSymbol i -> ProductSymbol j -> Entries i j (Entries i j x) -> Entries i j x

joining entries of entries.

Property Let xs' = etsJoin r c xs

Pre

For all (xij,i,j) in xs holds:

i < lengthN r and j < lengthN c

For all (_,i',j') in xij holds: i' < ri and j' < cj where ..ri.. = r, ..cj.. = c.

Post

xs' is valid.

etscr :: Eq i => Entries i j x -> Col i (Row j x)

the underlying column of rows.

etsrc :: (Ord i, Ord j) => Entries i j x -> Row j (Col i x)

the underlying row of columns.

crets :: Col i (Row j x) -> Entries i j x

the entries given by a column of rows.

rcets :: (Ord i, Ord j) => Row j (Col i x) -> Entries i j x

the entries given by a row of columns.

etsElimZeros :: Additive x => Entries i j x -> Entries i j x

elimination of zeros.

Row

newtype Row j x

viewing a partial sequence as a row.

Constructors

Row (PSequence j x)

rowxs :: Row j x -> [(x, j)]

underlying list of indexed entries.

rowEmpty :: Row j x

the empty row.

rowIsEmpty :: Row j x -> Bool

check for being empty.

rowHead :: Row j x -> (x, j)

head.

rowTail :: Row j x -> Row j x

tail.

rowFilter :: (x -> Bool) -> Row j x -> Row j x

filtering a row by the given predicate.

rowMapShift :: Number j => j -> ((x, j) -> y) -> Row j x -> Row j y

mapping and shifting of a row.

rowAppend :: Row j x -> Row j x -> Row j x

appending a row.

Property Let zs = rowAppend xs ys where ..(x,l) = xs and (y,f).. = ys then holds:

If

l < f

Then

zs is valid.

rowInterlace :: Ord j => (x -> y -> z) -> (x -> z) -> (y -> z) -> Row j x -> Row j y -> Row j z

interlacing two rows.

rowElimZeros :: Additive a => Row i a -> Row i a

elimination of zeros.

rowSwap :: Ord j => j -> j -> Row j x -> Row j x

swapping two entries of a row.

Pre k < l.

rowAdd :: (Additive a, Ord j) => Row j a -> Row j a -> Row j a

adding two rows.

rowMltl :: Distributive a => a -> Row j a -> Row j a

multiplies each element of the row by the given factor from the left.

rowShear :: Ord j => (Maybe x -> s -> Maybe x) -> (Maybe x -> Maybe x -> Maybe x) -> j -> j -> s -> s -> s -> s -> Row j x -> Row j x

shears two entries of a row.

Property Let r' = rowShear (<*) (+) k l s t u v r, then holds:

Pre

k < l.

Note rowShear is like multiplying the given row from the right with the matrix given by k l s t u v.

rowScale :: Ord j => (x -> s -> Maybe x) -> j -> s -> Row j x -> Row j x

scales the entry at the given position by the given factor.

Col

newtype Col i x

viewing a partial sequence as a column.

Constructors

Col (PSequence i x)

colxs :: Col i x -> [(x, i)]

underlying list of indexed entries.

colEmpty :: Col i x

the empty column.

colIsEmpty :: Col i x -> Bool

check for being empty.

colHead :: Col i x -> (x, i)

head.

colTail :: Col i x -> Col i x

tail.

colFilter :: (x -> Bool) -> Col i x -> Col i x

filtering a column by the given predicate.

colMapShift :: Number i => i -> ((x, i) -> y) -> Col i x -> Col i y

mapping and shifting of a column.

colAppend :: Col i x -> Col i x -> Col i x

appending a column..

Property Let zs = colAppend xs ys where ..(x,l) = xs and (y,f).. = ys then holds:

If

l < f

Then

zs is valid.

colInterlace :: Ord i => (x -> y -> z) -> (x -> z) -> (y -> z) -> Col i x -> Col i y -> Col i z

interlacing two columns.

colElimZeros :: Additive a => Col i a -> Col i a

elimination of zeros.

colSwap :: Ord i => i -> i -> Col i x -> Col i x

swapping two entries of a column.

Pre k < l.

colAdd :: (Additive a, Ord i) => Col i a -> Col i a -> Col i a

adding two columns.

colMltr :: Distributive a => Col i a -> a -> Col i a

multiplies each element of the column by the given factor from the right.

colShear :: Ord i => (s -> Maybe x -> Maybe x) -> (Maybe x -> Maybe x -> Maybe x) -> i -> i -> s -> s -> s -> s -> Col i x -> Col i x

shears two entries of a column.

Property Let c' = colShear (<*) (+) k l s t u v c, then holds:

Pre

k < l.

Note colShear is like multiplying the given column from the left with the matrix given by k l s t u v.

colScale :: Ord i => (s -> x -> Maybe x) -> i -> s -> Col i x -> Col i x

scales the entry at the given position by the given factor.

Col Row

crHeadColAt :: Eq j => j -> Col i (Row j a) -> Col i a

get the head column at j.

Pre for all j' in rws holds: j <= j'.

crHeadRowAt :: Eq i => i -> Col i (Row j a) -> Row j a

get the head row at i.

Pre for all i' in rws holdst: i <= i'.

Duality

coEntries :: (Ord i, Ord j) => Entries i j x -> Dual (Entries i j x)

to the dual of Entries, with inverse coEntriesInv.

coEntriesInv :: (Ord i, Ord j) => Dual (Entries i j x) -> Entries i j x

from the dual of Entries, with inverse coEntries.