OAlg.Entity.Matrix.Transformation

Description

elementary matrix transformations, i.e. operations of GLT on Matrix.

Row Trafo

newtype RowTrafo a

GLT as row transformations.

Constructors

RowTrafo (GLT a)

crTrafoRows :: Transformation x -> Col N (Row N x) -> Col N (Row N x)

applying a transformation as a row transformation on a column of rows.

Col Trafo

newtype ColTrafo x

GLT as a column transformation.

Constructors

ColTrafo (GLT x)

crTrafoCols :: Col N (Row N x) -> Transformation x -> Col N (Row N x)

applying a transformation as a column transformation on a column of rows.

Diagonal Form

data DiagonalForm k

the result of transforming a matrix into a diagonal form.

Property Let DiagonalForm ds rt ct be in DiagonalForm k, then holds:

n <= lengthN (start rt) and n <= lengthN (end ct) where n = lengthN ds.

For all d in ds holds: not (isZero d).

Constructors

DiagonalForm [k] (RowTrafo k) (ColTrafo k)

dgfMatrix :: Distributive k => DiagonalForm k -> Matrix k

the resulting matrix by applying on the diagonal matrix the inverse of the given transformations.

newtype DiagonalFormStrictPositive k

predicate for diagonal forms with strict positive entries.

Property Let DiagonalFormStrictPositive (DiagonalForm ds _ _) be in DiagonalForm k, then holds: 0 < d for all d in ds.

Constructors

DiagonalFormStrictPositive (DiagonalForm k)