OAlg.Entity.Product.Definition

Description

defintion of free products over Oriented symbols with exponents in a Number.

Note On Oriented structures the canonical injection inj and projection prj are bijections between the valid entities of Path and Product N. This is not true betwenn Path and ProductForm N as

>>> prj (P 3 :^ 2 :: ProductForm N Q) :: Path Q
Path () [3,3]

and

>>> prj (P 3 :* P 3 :: ProductForm N Q) :: Path Q
Path () [3,3]

both map to the same Path! But

>>> let p = make (P 3) :: Product N Q in p * p == p ^ 2
True

Product

data Product r a

free product over Oriented symbols in a with exponents in a Integral r.

Definition A Product p is valid if and only if its underlying ProductForm pf is valid and pf is reduced, i.e. pf == reduce pf.

prLength :: Product N a -> N

number of primary factors where where all simple factors are expanded according to there exponent.

prFactor :: Product N a -> N -> a

the n-th primary factor where all simple factors are expanded according to there exponent.

prFactors :: Product N a -> [a]

the list of primary factors.

prwrd :: (Integral r, Oriented a) => Product r a -> Word r a

restriction of prfwrd.

nProduct :: (Hom Ort h, Multiplicative x) => h a x -> Product N a -> x

mapping a product with exponents in N into a Multiplicative structure applying a homomorphism between Oriented structures.

zProduct :: (Hom Ort h, Cayleyan x) => h a x -> Product Z a -> x

mapping a product with exponents in Z into a Cayleyan structure applying a homomorphism between Oriented structures.

prdMapTotal :: (Singleton (Point y), Oriented y, Integral r) => (x -> y) -> Product r x -> Product r y

mapping a product.

prFromOp :: Product r (Op a) -> Product r a

from Op symbols.

Property For every Oriented structure a and Integral r the resulting map prFromOp is a contravariant homomorphisms between Multiplicative structures.

Word

newtype Word r a

list of symbols in a together with an exponent in r.

Constructors

Word [(a, r)]

fromWord :: Word r a -> [(a, r)]

the underlying list of a's with their exponent.

prfwrd :: Integral r => ProductForm r a -> Word r a

transforming a ProductForm to its corresponding Word.

wrdprf :: Semiring r => Point a -> Word r a -> ProductForm r a

transforming a Word to it corresponding ProductForm.

Note the Point is needed for empty Words.

wrdPrfGroup :: (Eq a, Semiring r) => Word r a -> Rdc (Word r a)

reducing a Word by adding the exponents of consecutive equal symbols and eliminating symbols with zero exponents.

nFactorize :: N -> Word N N

factorization of a natural number to powers of primes. For 0 there will be thrown Undefined.

nFactorize'

Arguments

:: N
bound for the primes

-> N
a natural number

-> Word N N

factorization of a natural number to powers of primes smaller then the given bound. For 0 there will be thrown Undefined.

Form

data ProductForm r a

form for a free product over Oriented symbols in a with exponents in r.

Definition Let r be a Number. A ProductForm pf is valid if and only if orientation pf is valid (see definition below) and all its symbols x - where P x occurs in pf - are valid.

The orientation of pf is defined according:

 orientation pf = case pf of
   One p    -> one p
   P a      -> orientation a
   f :^ r   -> orientation f ^ r where (^) = power
   f :* g   -> orientation f * orientation g

Note Number is required for -1, 0 and 1 are not degenerated as in Z/2 or Z/1.

Constructors

One (Point a)

P a

(ProductForm r a) :^ r infixl 9

(ProductForm r a) :* (ProductForm r a) infixr 7

prfLength :: Number r => ProductForm r a -> N

length.

prfDepth :: ProductForm r a -> N

depth.

prfFactors :: ProductForm N a -> [a]

list of elementary factors.

nProductForm :: (Hom Ort h, Multiplicative x) => h a x -> ProductForm N a -> x

mapping a product form with exponents in N into a Multiplicative structure applying a homomorphism between Oriented structures.

zProductForm :: (Hom Ort h, Cayleyan x) => h a x -> ProductForm Z a -> x

mapping a product form with exponents in Z into a Cayleyan structure applying a homomorphism between Oriented structures.

prfInverse :: Number r => ProductForm r a -> ProductForm r a

formal inverse

Let p in ProductForm r a then:

Pre If p contains a factor P a then minusOne /= Nothing.

Post the formal inverse.

prfFromOp :: ProductForm r (Op a) -> ProductForm r a

from Op symbols.

prfMapTotal :: Singleton (Point y) => (x -> ProductForm r y) -> ProductForm r x -> ProductForm r y

mapping a product form

Reduction

prfReduce :: (Oriented a, Integral r) => ProductForm r a -> ProductForm r a

reducing a ProductForm according to prfReduceWith return.

prfReduceWith :: (Oriented a, Integral r) => (Word r a -> Rdc (Word r a)) -> ProductForm r a -> ProductForm r a

reduces a product form by the given reduction rules for words until no more reductions are applicable.

Operations

prfopr :: (x -> t -> x) -> x -> ProductForm N t -> x

applicative operation from the right.

prfopr' :: N -> (x -> t -> x) -> x -> ProductForm N t -> x

partially strict version of prfopr, i.e. every n-th application will be reduced to head normal form.

Let x' = prfopr' n op x p.

Pre 0 < n.

Post x' == prfopr op x p.

prfopl :: (t -> x -> x) -> ProductForm N t -> x -> x

applicative operation from the left.

prfopl' :: N -> (t -> x -> x) -> ProductForm N t -> x -> x

partially strict version of prfopl, i.e. every n-th application will be reduced to head normal form.

Let x' = prfopl' n op p x.

Pre 0 < n.

Post x' == prfopl op p x.