OAlg.Entity.Product.Proposition

Description

propositions on free products over Oriented symbols with exponents in a Number.

Proposition

prpProduct :: Statement

validity of Product r (Orientation Symbol) for r equal to N and Z respectively

prpOrtProductNOrntSymbol :: Statement

validity of Product N (Orientation Symbol) being Oriented.

prpOrtProductZOrntSymbol :: Statement

validity of Product Z (Orientation Symbol) being Oriented.

prpMltProductNOrntSymbol :: Statement

validity of Product N (Orientation Symbol) being Multiplicative.

prpMltProductZOrntSymbol :: Statement

validity of Product Z (Orientation Symbol) being Multiplicative.

Random Variables

xStartProduct :: (Oriented a, Integral r) => XOrtSite From a -> X r -> N -> XOrtSite From (Product r a)

random variable of products generated from product forms with a maximal given depth (:^ dose not increases the depth).

xStartProductForm :: (Oriented a, Number r) => XOrtSite From a -> X r -> N -> XOrtSite From (ProductForm r a)

random variable of product forms with maximal depth d (a :^ constructor dose not increases the depth).

xPrdSymStart :: Integral r => N -> X r -> XOrtSite From (Product r (Orientation Symbol))

the induced random variable on Orientation Symbol).

xPrdSymMlt :: Integral r => N -> X r -> XMlt (Product r (Orientation Symbol))

the induced random variable for validating Multiplicative structures.

Example

xT :: XOrtSite From (Orientation Char)

example of a XStart for the quiver having two points 'a' and 'b' and two arrows 'a':>'a' and 'a':>'b'.

dstT :: Int -> N -> IO ()

its distribution