OAlg.Entity.Sequence

Description

Sequences of indexed items in a type x with an index type i and there permutations. We consider two kinds of sequences:

Complete

Sequences x 0,x 1,x 2,.. with an integral index type i where there indices don't have wholes and start at 0, e.g. [x] or free products of symbols in x i.e. ProductSymbol.

Partial

Sequences x i0,x i1,x i2,.. with a totally ordered index type i where there indices allow wholes, e.g. PSequence, Set, Graph.

Furthermore there are total right operations of Permutation defined on them which permutes the corresponding indices to yield a new sequence.

Documentation

module OAlg.Entity.Sequence.Definition

module OAlg.Entity.Sequence.CSequence

module OAlg.Entity.Sequence.PSequence

module OAlg.Entity.Sequence.Set

module OAlg.Entity.Sequence.Graph

module OAlg.Entity.Sequence.Permutation