OAlg.Entity.Sequence.CSequence

Description

completely defined sequences of items in x with index type N.

Sequence

type CSequence = ProductSymbol

completely defined sequences of items, i.e. free products with index type N.

module OAlg.Entity.Product.ProductSymbol

X

xCSequence :: Entity x => N -> X x -> X (CSequence x)

random variable for comletely defined sequences with the given maximal length.