OAlg.Entity.Sequence.Definition

Description

basic definitions for sequences as mappings of an index to an entity.

Sequence

class (LengthN (s x), Ord i) => Sequence s i x where

sequences as mappings of an index.

Definition Let s, i, x be an instance of Sequence and xs be in s x, then we call xs finite if and only if the evaluation of lengthN xs terminates and will not end up in an exception.

Property Let s, i, x be an instance of Sequence, then holds:

For all xs in s x holds:

graph is constant in its first parameter.

If xs is finite, then lengthN xs == lengthN (graph p xs) for any p.

For all xs in s x holds:

list is constant in its first parameter.

For all ..(x,i):(x,j).. in xs holds: i < j.

If xs is finite, then lengthN xs == lengthN (list p xs) for any p.

Let xs be in s x and i in i, then holds: there exists an x in x with xs ? i matches Just x if and only if there exists an (i',x) in graph (Just i) xs such that i == i'.

Note The first parameter of graph - respectively list - serves only as a proxy and as such it is only relevant on the type level.

Minimal complete definition

graph | list

Methods

graph :: p i -> s x -> Graph i x

the associated graph of a sequence

list :: p i -> s x -> [(x, i)]

the associated list of its items together with there index.

(??) :: s x -> i -> Maybe x

the i-th item.

listN :: Sequence s N x => s x -> [(x, N)]

the indexed list of the sequence.

(?) :: Sequence s i x => s x -> i -> x

the i-th element of the sequence.

Property Let xs be in s x and i in i for a instance of Sequence s i x, then holds: If i is in the support of xs then xs ? i is the i-th item of xs, else its evaluation will end up by throwing a IndexOutOfSupport-exception.

isEmpty :: Sequence s i x => p i -> s x -> Bool

checks for being empty.

span :: Sequence s i x => p i -> s x -> Span i

the span of a sequence.

support :: Sequence s i x => p i -> s x -> Set i

the support of a sequence, i.e. all the indices which are not mapped to Nothing.

image :: (Sequence s i x, Ord x) => p i -> s x -> Set x

the image of a sequence, i.e. all the entities are hit by the mapping.

Constructable

class (Entity x, Entity i, Sequence s i x) => ConstructableSequence s i x where

constructable sequences.

Minimal complete definition

sequence

Methods

sequence :: (i -> Maybe x) -> Set i -> s x

constructs a sequence.

(<&) :: s x -> Set i -> s x infixl 7

restricts a sequence.

sqcIndexMap :: (ConstructableSequence s i x, Sequence s j x) => Set i -> (i -> j) -> s x -> s x

mapping the indices according to the given set.

Exception

data SequenceException

sequence exceptions which are sub exceptions from SomeOAlgException.

Constructors

IndexOutOfSupport