OAlg.Entity.Slice.Free

Description

sliced structures with an assigned free Point of some given dimension.

Free

newtype Free k c

index for free points within a Multiplicative structure c.

>>> lengthN (Free attest :: Free N3 c)
3

Constructors

Free (Any k)

freeN :: Free k c -> N

the underlying natural number.

castFree :: Free k x -> Free k y

casting between Free types.

isFree :: (Eq (Point c), Sliced (Free k) c) => Free k c -> Point c -> Bool

check for being a free point, i.e. if it is equal to slicePoint.

Definition Let n be in Free n c and p in Point c then we call p of order n if and only if slicePoint i == p.

data SomeFree c where

some free attest.

Constructors

SomeFree :: (Attestable k, Sliced (Free k) c) => Free k c -> SomeFree c

data SomeFreeSlice s c where

some free point within a Multiplicative structure c.

Constructors

SomeFreeSlice :: (Attestable k, Sliced (Free k) c) => Slice s (Free k) c -> SomeFreeSlice s c

Limes

data LimesFree s p t n m a where

predicate for a limes with a free tip of its universal cone.

Property Let 'LimesFree k l be in LimesFree s p t n m a and then holds: slicePoint k == t where t = tip (universalCone l).

Constructors

LimesFree :: (Attestable k, Sliced (Free k) a) => Free k a -> Limes s p t n m a -> LimesFree s p t n m a

limesFree :: LimesFree s p t n m a -> Limes s p t n m a

the underlying free limes.

data DiagramFree t n m a

predicate for diagrams with free points.

Constructors

DiagramFree (FinList n (SomeFree a)) (Diagram t n m a)

dgfDiagram :: DiagramFree t n m a -> Diagram t n m a

the underlying diagram.

data KernelSliceFromSomeFreeTip n i c where

predicate for a kernel with a start point of its diagram given by the slice index and a free universal tip.

Property Let KernelSliceFromSomeFreeTip k' i ker be in KernelSliceFromSomeFreeTip n c, then holds:

slicePoint i == s where DiagramParallelLR s _ _ = diagram ker.

slicePoint k' == tip (universalCone ker).

Constructors

KernelSliceFromSomeFreeTip :: (Attestable k', Sliced (Free k') c) => Free k' c -> i c -> Kernel n c -> KernelSliceFromSomeFreeTip n i c

ksfKernel :: KernelSliceFromSomeFreeTip n i c -> Kernel n c

the underlying kernel.

Kernel

type KernelFree = LimesFree Dst Projective (Parallel LeftToRight) N2

kernel of a diagram with free points.

type KernelDiagramFree = DiagramFree (Parallel LeftToRight) N2

kerne diagram with free points.

Cokernel

type CokernelDiagramFree = DiagramFree (Parallel RightToLeft) N2

cokernel diagrams with free points.

Pullback

type PullbackFree n c = LimesFree Mlt Projective (Star To) (n + 1) n c

pullback of a diagram with free points.

type PullbackDiagramFree n c = DiagramFree (Star To) (n + 1) n c

pullback diagram with free points.