OAlg.Entity.Sum.Definition

Description

definition of free Sums over Fibred symbols.

Sum

data Sum r a

free sum over Fibred symbols in a with scalars in r.

Definition A Sum s is valid if and only if its underlying SumForm s' is valid and s' is reduced, i.e. s' == reduce s'.

smlc :: Semiring r => Sum r a -> LinearCombination r a

the associated linear combination.

Note The associated linear combination of a sum is sorted according to the second component!

smJoin :: (Semiring r, Commutative r, Fibred a, Ord a) => Sum r (Sum r a) -> Sum r a

joining a sum of sums.

nSum :: (Hom Fbr h, Additive x) => h a x -> Sum N a -> x

additive homomorphism for sums over N.

zSum :: (Hom Fbr h, Abelian x) => h a x -> Sum Z a -> x

additive homomorphism for sums over Z.

smMap :: (Singleton (Root y), Fibred y, Ord y, Semiring r, Commutative r) => (x -> y) -> Sum r x -> Sum r y

additive homomorphism to a totally defined sum.

Form

data SumForm r a

form for a free sum over Fibred symbols in a with scalars in r.

Definition Let r be a Commutative Semiring and a a Fibred structure. A SumForm a is valid if and only if all scalars in a are valid and all symbols in a are valid and have the same root.

Constructors

Zero (Root a)

S a

r :! (SumForm r a) infixr 9

(SumForm r a) :+ (SumForm r a) infixr 6

smfLength :: Number r => SumForm r a -> N

the length of a sum form,

smflc :: Semiring r => SumForm r a -> LinearCombination r a

transforming a sum form to its corresponding linear combination..

lcsmf :: Semiring r => Root a -> LinearCombination r a -> SumForm r a

transforming a word to its corresponding sum form.

smfMap :: Singleton (Root y) => (x -> y) -> SumForm r x -> SumForm r y

mapping of sum forms.

smfJoin :: SumForm r (SumForm r a) -> SumForm r a

joining a sum form of sum forms.

smfReduce :: (Fibred a, Ord a, Semiring r, Commutative r) => SumForm r a -> SumForm r a

reducing a sum form to its canonical form,

Linear Combination

newtype LinearCombination r a

list of symbols in a together with a scalar in r.

Note valid linear combinations must not be sorted according to the second component!

Constructors

LinearCombination [(r, a)]

lcs :: LinearCombination r a -> [(r, a)]

the underlying list of symbols with their scalar.

lcAggr :: (Eq a, Semiring r) => LinearCombination r a -> LinearCombination r a

aggregating linear combinations with same symbols.

lcSort :: Ord a => LinearCombination r a -> LinearCombination r a

sorting a linear combination according to its symbols.

lcSclFilter :: (r -> Bool) -> LinearCombination r a -> LinearCombination r a

filtering a word according to the scalars.