OAlg.Entity.Sum.SumSymbol

Description

free sums with symbols in a.

SumSymbol

newtype SumSymbol r a

free sum with symbols in a.

Constructors

SumSymbol (Sum r (R a))

ssypsq :: Semiring r => SumSymbol r a -> PSequence a r

the underlying partial sequence.

ssylc :: Semiring r => SumSymbol r a -> LinearCombination r a

the underlying linear combination.

sumSymbol :: (Semiring r, Commutative r, Entity a, Ord a) => [(r, a)] -> SumSymbol r a

the induced free sum given by a list of scalars and symbols.

sy :: (Semiring r, Commutative r, Entity a, Ord a) => a -> SumSymbol r a

the induced free sum given by the symbol.

ssyMap :: (Semiring r, Commutative r, Entity y, Ord y) => (x -> y) -> SumSymbol r x -> SumSymbol r y

mapping of free sums

ssySum :: (Semiring r, Commutative r, Entity y, Ord y) => (x -> LinearCombination r y) -> SumSymbol r x -> SumSymbol r y

additive homomorphism given by a mapping of a symbol in x to a linear combination of y.

ssyJoin :: (Semiring r, Commutative r, Entity x, Ord x) => SumSymbol r (SumSymbol r x) -> SumSymbol r x

joining a free sum of free sums to a free sum.

ssyprj :: (Semiring r, Commutative r, Ord a, Entity a) => Set a -> SumSymbol r a -> SumSymbol r a

the projectin of a free sum according to the given set of symbols.

Definition Let x be in SumSymbol r a and s a Set of symbols in a, then x is called representable according to s iff all symbols of ssylc x are elements of s.

Property Let s be a set of symbols in a and x be representable in SumSymbol r a according to s, then ssyprj x == x.

Examples

>>> ssyprj (Set [A,D,E]) (3!sy D) :: SumSymbol Z Symbol
SumSymbol[3!D]

>>> ssyprj (Set [A,D,E]) (2!sy B) :: SumSymbol Z Symbol
SumSymbol[]

>>> ssyprj (Set [A,D,E]) (3!sy D + sy A - 5!sy E) :: SumSymbol Z Symbol
SumSymbol[A+3!D+-5!E]

>>> ssyprj (Set [A,D,E]) (2!sy D + 7!sy B - sy E + sy F) :: SumSymbol Z Symbol
SumSymbol[2!D+-1!E]

R

newtype R a

adjoining the root ().

Constructors

R a