OAlg.Hom.Multiplicative.Definition

Description

definition of homomorphisms between Multiplicative structures.

Multiplicative

class (EmbeddableMorphism h Mlt, HomOriented h) => HomMultiplicative h

type family of homomorphisms between Multiplicative structures.

Propoerty Let h be a type instance of the class HomMultiplicative, then for all a, b and f in h a b holds:

For all p in Point a holds: amap f (one p) == one (pmap f p).

For all x, y in a with start x == end y holds: amap f (x * y) == amap f x * amap f y.

Such a h will be called a family of homomorphisms between multiplicative structures and an entity f of h a b a multiplicative homomorphism between a and b.

Note If we interpret the types a and b as small categories (see note at Multiplicative) then we can interpret the type family h as a family of covariant functors.

type IsoMultiplicative h = (FunctorialHomOriented h, Cayleyan2 h, HomMultiplicative h)

isomorphisms between Multiplicative structures.

OpHom

isoFromOpOpMlt :: Multiplicative a => IsoOp Mlt (Op (Op a)) a

the induced isomorphism of Multiplicative structures given by FromOpOp.

isoOppositeMlt :: Entity p => IsoOp Mlt (Op (Orientation p)) (Orientation p)

the induced isomorphism of Oriented structures given by Opposite.