OAlg.Hom.Multiplicative.Proposition

Description

propositions on homomorphisms between Multiplicative structures.

Proposition

prpHomOpMlt :: Statement

validity of HomOp Mlt to be a family of Multiplicative homomorphisms.

Multiplicative

prpHomMlt :: Hom Mlt h => XHomMlt h -> Statement

validity of homomorphisms between Multiplicative structures according to OAlg.Hom.Multiplicative.

data XHomMlt h

random variable for validating a type family h according to HomMultiplicative.

Constructors

XHomMlt (X (SomeApplPnt h)) (X (SomeApplMltp2 h))

coXHomMlt :: HomMultiplicative h => XHomMlt h -> Dual (XHomMlt h)

to the dual of XHomMlt h with its inverse coXHomMltInv.

coXHomMltInv :: HomMultiplicative h => Dual (XHomMlt h) -> XHomMlt h

from the dual of Dual (XHomMlt h) with its inverse coXHomMlt.

data SomeApplPnt h where

some application on a point.

Constructors

SomeApplPnt :: h a b -> Point a -> SomeApplPnt h

coSomeApplPnt :: Transformable1 Op (ObjectClass h) => SomeApplPnt h -> Dual (SomeApplPnt h)

to the dual of SomeApplPnt h with its inverse coSomeApplPntInv.

coSomeApplPntInv :: Dual (SomeApplPnt h) -> SomeApplPnt h

from the dual of Dual (SomeApplPnt h) with its inverse coSomeApplPnt.

data SomeApplMltp2 h where

some application on multiplicable factors.

Constructors

SomeApplMltp2 :: h a b -> Mltp2 a -> SomeApplMltp2 h

coSomeApplMltp2 :: HomMultiplicative h => SomeApplMltp2 h -> Dual (SomeApplMltp2 h)

to the dual of SomeApplMltp2 h with its inverse coSomeApplMltp2Inv.

coSomeApplMltp2Inv :: HomMultiplicative h => Dual (SomeApplMltp2 h) -> SomeApplMltp2 h

from the dual of Dual (SomeApplMltp2 h) with its inverse coSomeApplMltp2.

prpHomMlt1 :: Hom Mlt h => h a b -> Point a -> Statement

validity according to OAlg.Hom.Multiplicative.

prpHomMlt2 :: Hom Mlt h => h a b -> Mltp2 a -> Statement

validity according to OAlg.Hom.Multiplicative.

X

xHomMlt :: Hom Mlt h => SomeApplMlt d h -> XHomMlt h

the induced random variable of XHomMlt, given by SomeApplMlt.

xHomMlt' :: h a b -> XMlt a -> XHomMlt h

the induced random variable of XHomMlt.

xSomeApplPnt :: SomeApplMlt d h -> X (SomeApplPnt h)

random variable for some application on a point given by a SomeApplMlt.

xSomeApplMltp2 :: Hom Mlt h => SomeApplMlt d h -> X (SomeApplMltp2 h)

random variable for some application on multiplicable factors given by a SomeApplMlt.