OAlg.Hom.Oriented.Proposition

Description

propositions on homomorphisms between Oriented structures.

Proposition

prpIdHomOrt :: Statement

validity of IdHom Ort to be a family of Oriented homomorphisms between Orientation Symbol and Z.

prpHomOpOrt :: Statement

validity of HomOp Ort according to HomOriented on Orientation Symbol.

prpIsoOpOrtCategory :: Statement

validity of IsoOp Ort according to Category on Orientation Symbol.

prpIsoOpOrtFunctorial :: Statement

validity of IsoOp Ort according Functorial.

Oriented

prpHomOrt :: Hom Ort h => XHomOrt h -> Statement

validity of homomorphisms between Oriented structures based on the given random variable.

type XHomOrt h = XAppl h

random variable to validate homomorphisms between Oriented structures.

prpHomOrt' :: Hom Ort h => h a b -> XOrt a -> Statement

validity of homomorphisms between Oriented structures based on the given random variable.

prpHomOrt1 :: Hom Ort h => h a b -> a -> Statement

validity of homomorphisms between Oriented structures based on the given values.

relHomOrtHomomorphous :: Hom Ort h => Homomorphous Ort a b -> h a b -> a -> Statement

validity of homomorphisms between Oriented for a given value in the domain.

X

xIsoOpOrtFrom :: XFnctMrphSite From (IsoOp Ort)

random variale of IsoOp Ort.