OAlg.Limes.Cone.Definition

Description

definition of Cones over Diagrams.

Cone

data Cone s p t n m a where

cone over a diagram.

Properties Let c be in Cone s p t n m a then holds:

If c matches ConeProjective d t cs for a Multiplicative structure a then holds:

For all ci in cs holds: start ci == t and end ci == pi where pi in dgPoints d.

For all aij in dgArrows d holds: aij * ci == cj where ci, cj in cs.

If c matches ConeInjective d t cs for a Multiplicative structure a then holds:

For all ci in cs holds: end ci == t and start ci == pi where pi in dgPoints d.

For all aij in dgArrows d holds: cj * aij == ci where ci, cj in cs.

If c matches ConeKernel p k for a Distributive structure a then holds:

end k == p0 where p0 in dgPoints p

For all a in dgArrows p holds: a * k == zero (t :> p1) where t = start k and p0, p1 in dgPoints p.

If c matches ConeCokernel p k for a Distributive structure a then holds:

start k == p0 where p0 in cnPoints c

For all a in dgArrows p holds: k * a == zero (p1 :> t) where t = end k and p0, p1 in dgPoints p.

Constructors

ConeProjective :: Multiplicative a => Diagram t n m a -> Point a -> FinList n a -> Cone Mlt Projective t n m a

ConeInjective :: Multiplicative a => Diagram t n m a -> Point a -> FinList n a -> Cone Mlt Injective t n m a

ConeKernel :: Distributive a => Diagram (Parallel LeftToRight) N2 m a -> a -> Cone Dst Projective (Parallel LeftToRight) N2 m a

ConeCokernel :: Distributive a => Diagram (Parallel RightToLeft) N2 m a -> a -> Cone Dst Injective (Parallel RightToLeft) N2 m a

data Perspective

concept of Projective and Injective.

Constructors

Projective

Injective

cnDiagram :: Cone s p t n m a -> Diagram t n m a

the underlying diagram.

cnDiagramTypeRefl :: Cone s p t n m a -> Dual (Dual t) :~: t

reflexivity of the underlying diagram type.

tip :: Cone s p t n m a -> Point a

the tip of a cone.

Property Let c be in Cone s p t n m a for a Oriented structure then holds:

If p is equal to Projective then holds: start ci == tip c for all ci in shell c.

If p is equal to Injective then holds: end ci == tip c for all ci in shell c.

shell :: Cone s p t n m a -> FinList n a

the shell of a cone.

Property Let c be in Cone s p t n m a for a Oriented structure then holds:

If p is equal to Projective then holds: fmap end (shell c) == cnPoints c.

If p is equal to Injective then holds: fmap start (shell c) == cnPoints c.

cnArrows :: Cone s p t n m a -> FinList m a

the arrows of the underlying diagram, i.e. dgArrows . cnDiagram.

cnPoints :: Oriented a => Cone s p t n m a -> FinList n (Point a)

the points of the underlying diagram, i.e. dgPoints . cnDiagram.

cnMap :: Hom s h => h a b -> Cone s p t n m a -> Cone s p t n m b

mapping of a cone.

cnMapMlt :: Hom Mlt h => h a b -> Cone Mlt p t n m a -> Cone Mlt p t n m b

mapping of a cone under a Multiplicative homomorphism.

cnMapDst :: Hom Dst h => h a b -> Cone Dst p t n m a -> Cone Dst p t n m b

mapping of a cone under a Distributive homomorphism.

Distributive

cnZeroHead :: Cone Dst p t n m a -> ConeZeroHead Mlt p t n (m + 1) a

embedding of a cone in a distributive structure to its multiplicative cone.

cnKernel :: (Distributive a, p ~ Projective, t ~ Parallel LeftToRight) => ConeZeroHead Mlt p t n (m + 1) a -> Cone Dst p t n m a

the kernel cone of a zero headed parallel cone, i.e. the inverse of cnZeroHead.

cnCokernel :: (Distributive a, p ~ Injective, t ~ Parallel RightToLeft) => ConeZeroHead Mlt p t n (m + 1) a -> Cone Dst p t n m a

the cokernel cone of a zero headed parallel cone, i.e. the inverse of cnZeroHead.

cnDiffHead :: (Distributive a, Abelian a) => Cone Mlt p (Parallel d) n (m + 1) a -> ConeZeroHead Mlt p (Parallel d) n (m + 1) a

subtracts to every arrow of the underlying parallel diagram the first arrow and adapts the shell accordingly.

newtype ConeZeroHead s p t n m a

predicate for cones where the first arrow of its underlying diagram is equal to zero.

Constructors

ConeZeroHead (Cone s p t n m a)

coConeZeroHead :: ConeZeroHead s p t n m a -> Dual (ConeZeroHead s p t n m a)

to the dual, with its inverse coConeZeroHead.

czFromOpOp :: ConeStruct s a -> ConeZeroHead s p t n m (Op (Op a)) -> ConeZeroHead s p t n m a

from the bidual.

coConeZeroHeadInv :: ConeStruct s a -> (Dual (Dual p) :~: p) -> (Dual (Dual t) :~: t) -> Dual (ConeZeroHead s p t n m a) -> ConeZeroHead s p t n m a

from the dual, with its inverse coConeZeroHead.

Duality

cnToOp :: ConeDuality s f g a -> f a -> g (Op a)

to g (Op a).

cnFromOp :: ConeDuality s f g a -> g (Op a) -> f a

from g (Op a).

data ConeDuality s f g a where

Op-duality between cone types.

Constructors

ConeDuality :: ConeStruct s a -> (f a :~: Cone s p t n m a) -> (g (Op a) :~: Dual (Cone s p t n m a)) -> (Dual (Dual p) :~: p) -> (Dual (Dual t) :~: t) -> ConeDuality s f g a

coCone :: Cone s p t n m a -> Dual (Cone s p t n m a)

to the dual cone, with its inverse coConeInv.

coConeInv :: ConeStruct s a -> (Dual (Dual p) :~: p) -> (Dual (Dual t) :~: t) -> Dual (Cone s p t n m a) -> Cone s p t n m a

from the dual cone, with its inverse coCone.

cnFromOpOp :: ConeStruct s a -> Cone s p t n m (Op (Op a)) -> Cone s p t n m a

from Op . Op.

Cone Struct

data ConeStruct s a where

cone structures.

Constructors

ConeStructMlt :: Multiplicative a => ConeStruct Mlt a

ConeStructDst :: Distributive a => ConeStruct Dst a

cnStructOp :: ConeStruct s a -> ConeStruct s (Op a)

the opposite cone structure.

cnStructMlt :: ConeStruct s a -> Struct Mlt a

the Multiplicative structure of a cone structure.

cnStruct :: ConeStruct s a -> Struct s a

the associated structure of a cone structure.

Orientation

cnPrjOrnt :: Entity p => p -> Diagram t n m (Orientation p) -> Cone Mlt Projective t n m (Orientation p)

the projective cone on Orientation with the underlying given diagram and tip with the given point.

cnInjOrnt :: Entity p => p -> Diagram t n m (Orientation p) -> Cone Mlt Injective t n m (Orientation p)

the injective cone on Orientation with the underlying given diagram and tip with the given point.

Chain

cnPrjChainTo :: Multiplicative a => FactorChain To n a -> Cone Mlt Projective (Chain To) (n + 1) n a

the induced Projective cone with ending factor given by the given FactorChain.

Property Let h = FactorChain f d be in FactorChain To n a for a Multiplicative structure a and ConeProjective d' _ (_:|..:|c:|Nil) = cnPrjChainTo h then holds: d' == d and c == f.

cnPrjChainToInv :: Cone Mlt Projective (Chain To) (n + 1) n a -> FactorChain To n a

the underlying factor chain of a projective chain to cone, i.e the inverse of cnPrjChainToInv.

cnPrjChainFrom :: Multiplicative a => FactorChain From n a -> Cone Mlt Projective (Chain From) (n + 1) n a

the induced Projective cone with starting factor given by the given FactorChain.

Property Let h = FactorChain f d be in FactorChain From n a for a Multiplicative structure a and ConeProjective d' _ (c:|_) = cnPrjChainFrom h then holds: d' == d and c == f.

cnPrjChainFromInv :: Cone Mlt Projective (Chain From) (n + 1) n a -> FactorChain From n a

the underlying factor chain of a projective chain from cone, i.e. the inverse of cnPrjChainFrom.

data FactorChain s n a

predicate for a factor with end point at the starting point of the given chain.

Property

Let FactorChain f d be in FactorChain To n a for a Multiplicative structure a then holds: end f == chnToStart d.

Let FactorChain f d be in FactorChain From n a for a Multiplicative structure a then holds: end f == chnFromStart d.

Constructors

FactorChain a (Diagram (Chain s) (n + 1) n a)