OAlg.Limes.Definition

Description

definition of a Limes of a Diagram.

Limes

data Limes s p t n m a where

limes of a diagram, i.e. a distinguished cone over a given diagram having the following universal property

Property Let a be a Multiplicative structure and u in Limes s p t n m a then holds: Let l = universalCone u in

l is valid.

eligibleCone u l.

eligibleFactor u l (one (tip l)).

For all c in Cone s p t n m a with eligibleCone u c holds: Let f = universalFactor u c in

f is valid.

eligibleFactor u c f.

For all x in a with eligibleFactor u c x holds: x == f.

Note

As the function universalFactor l for a given limes l is uniquely determined by the underlying cone of l, it is permissible to test equality of two limits just by there underling cones. In every computation equal limits can be safely replaced by each other.

It is not required that the evaluation of universal factor on a non eligible cone yield an exception! The implementation of the general algorithms for limits do not check for eligibility.

Constructors

LimesProjective :: Cone s Projective t n m a -> (Cone s Projective t n m a -> a) -> Limes s Projective t n m a

LimesInjective :: Cone s Injective t n m a -> (Cone s Injective t n m a -> a) -> Limes s Injective t n m a

universalPoint :: Limes s p t n m a -> Point a

the universal point of a limes, i.e. the tip of the universal cone.

universalCone :: Limes s p t n m a -> Cone s p t n m a

the underlying universal cone of a limes.

universalShell :: Limes s p t n m a -> FinList n a

the shell of the universal cone.

universalFactor :: Limes s p t n m a -> Cone s p t n m a -> a

the universal factor of a Limes l to a given eligible cone.

Property Let l be in Limes s p t n m a then holds: For all c in Cone s p t n m a with eligibleCone l c holds: eligibleFactor l c (universalFactor l c).

diagram :: Limes s p t n m a -> Diagram t n m a

the underlying diagram of a limes.

lmDiagramTypeRefl :: Limes s p t n m a -> Dual (Dual t) :~: t

reflexivity of the underlying diagram type.

eligibleCone :: Oriented a => Limes s p t n m a -> Cone s p t n m a -> Bool

eligibility of a cone with respect to a limes.

Property Let u be in Limes s p t n m a and c in Cone s p t n m a then holds: eligibleCone u c is true if and only if diagram u == cnDiagram c is true.

eligibleFactor :: Limes s p t n m a -> Cone s p t n m a -> a -> Bool

eligibility of a factor with respect to a limes and a cone.

Property Let u be in Limes s p t n m a, c in Cone s p t n m a with eligibleCone u c and x in a then holds:

If u matches LimesProjective l _ then holds: eligibleFactor u c x is true if and only if cnEligibleFactor x c l is true.

If u matches LimesInjective l _ then holds: eligibleFactor u c x is true if and only if cnEligibleFactor x l c is true.

lmMap :: IsoOrt s h => h a b -> Limes s p t n m a -> Limes s p t n m b

mapping between limits.

Duality

lmToOp :: LimesDuality s f g a -> f a -> g (Op a)

to g (Op a).

lmFromOp :: LimesDuality s f g a -> g (Op a) -> f a

from g (Op a).

data LimesDuality s f g a where

Op-duality between limes types.

Constructors

LimesDuality :: ConeStruct s a -> (f a :~: Limes s p t n m a) -> (g (Op a) :~: Dual (Limes s p t n m a)) -> (Dual (Dual p) :~: p) -> (Dual (Dual t) :~: t) -> LimesDuality s f g a

coLimes :: ConeStruct s a -> (Dual (Dual p) :~: p) -> (Dual (Dual t) :~: t) -> Limes s p t n m a -> Dual (Limes s p t n m a)

the co limes with its inverse coLimesInv.

coLimesInv :: ConeStruct s a -> (Dual (Dual p) :~: p) -> (Dual (Dual t) :~: t) -> Dual (Limes s p t n m a) -> Limes s p t n m a

the inverse of coLimes.

lmFromOpOp :: ConeStruct s a -> (Dual (Dual p) :~: p) -> (Dual (Dual t) :~: t) -> Limes s p t n m (Op (Op a)) -> Limes s p t n m a

from Op . Op.

Construction

lmToPrjOrnt :: (Entity p, a ~ Orientation p) => p -> Diagram t n m a -> Limes Mlt Projective t n m a

projective limes on oriented structures.

lmFromInjOrnt :: (Entity p, a ~ Orientation p) => p -> Diagram t n m a -> Limes Mlt Injective t n m a

injective limes on oriented structures.

Proposition

relLimes :: ConeStruct s a -> XOrtPerspective p a -> Limes s p t n m a -> Statement

validity of a Limes.

Exception

data LimesException

limes exceptions which are sub exceptions from SomeOAlgException.

Constructors

ConeNotEligible String