OAlg.Prelude

Description

To avoid ambiguity for the algebraic operators on should exclude the standard Prelude and use this one instead.

Category

module OAlg.Category.Applicative

module OAlg.Category.Definition

module OAlg.Category.Proposition

Validating

module OAlg.Data.Statement

module OAlg.Data.X

module OAlg.Data.Validable

module OAlg.Control.Validate

Data

module OAlg.Data.Boolean

module OAlg.Data.Equal

module OAlg.Data.Maybe

module OAlg.Data.Show

module OAlg.Data.Number

Entity

module OAlg.Entity.Definition

Structure

module OAlg.Structure.Definition

Dual

module OAlg.Data.Dualisable

module OAlg.Data.Opposite

Ord

module OAlg.Data.Ord

Additionals

Some additional definition from the standard Prelude.

Type Equality

class a ~# b => (a :: k) ~ (b :: k)

Bounded

class Bounded a where

Methods

minBound :: a

maxBound :: a

IO

data IO a

putStrLn :: String -> IO ()

seq

seq :: forall {r :: RuntimeRep} a (b :: TYPE r). a -> b -> b

Undefined

undefined :: forall (r :: RuntimeRep) (a :: TYPE r). HasCallStack => a

error :: forall (r :: RuntimeRep) (a :: TYPE r). HasCallStack => [Char] -> a

Exception

module OAlg.Control.Exception