OAlg.Structure.Additive.Proposition

Description

propositions on Additive structures.

Additive

relIsZero :: Additive a => a -> Statement

relation of isZero.

prpAdd :: Additive a => XAdd a -> Statement

validity of the Additive structure of a.

Note For a good choice of X N see the note of prpAdd6.

prpAdd1 :: Additive a => p a -> X (Root a) -> Statement

validtiy of root according to OAlg.Structure.Additive.Definition.

prpAdd2 :: Additive a => X (a, a) -> Statement

validtiy of two summands according to OAlg.Structure.Additive.Definition.

prpAdd2_1 :: Additive a => Adbl2 a -> Statement

validity of + for two addable summands according to OAlg.Structure.Additive.Definition.

prpAdd2_2 :: Additive a => a -> a -> Statement

validity of + for two not addable summands according to OAlg.Structure.Additive.Definition.

prpAdd3 :: Additive a => X (Adbl2 a) -> Statement

validtiy of two adable summands according to OAlg.Structure.Additive.Definition.

prpAdd4 :: Additive a => X a -> Statement

validtiy according to OAlg.Structure.Additive.Definition.

prpAdd5 :: Additive a => X (Adbl3 a) -> Statement

validtiy of three adable summands according to OAlg.Structure.Additive.Definition.

prpAdd6 :: Additive a => X N -> X a -> Statement

validity of ntimes according to OAlg.Structure.Additive.Definition.

Abelian

prpAbl :: Abelian a => XAbl a -> Statement

validity of the Abelian structure of a.

prpAbl1 :: Abelian a => X a -> Statement

validity according to OAlg.Structure.Additive.Definition.

prpAbl2 :: Abelian a => X a -> Statement

validity according to OAlg.Structure.Additive.Definition.

prpAbl3 :: Abelian a => X (a, a) -> Statement

validity of two summands according to OAlg.Structure.Additive.Definition.

prpAbl3_1 :: Abelian a => Adbl2 a -> Statement

validity of - for two addable summands according to OAlg.Structure.Additive.Definition.

prpAbl3_2 :: Abelian a => a -> a -> Statement

validity of - for two not addable summands according to OAlg.Structure.Additive.Definition.

prpAbl4 :: Abelian a => X (Adbl2 a) -> Statement

validity of two summands according to OAlg.Structure.Additive.Definition.

prpAbl5 :: Abelian a => X Z -> X a -> Statement

validity of two summands according to OAlg.Structure.Additive.Definition.

X

Additive

data XAdd a

random variable for validating Additive structures.

Constructors

XAdd (X N) (X (Root a)) (X a) (X (Adbl2 a)) (X (Adbl3 a))

data Adbl2 a

predicate for two addable summands.

Constructors

Adbl2 a a

data Adbl3 a

predicate for three addable summands.

Constructors

Adbl3 a a a

class XStandardAdd a where

standard random variable for Additive structures.

Methods

xStandardAdd :: XAdd a

Abelian

data XAbl a

random variable for validating Abelian structures.

Constructors

XAbl (X Z) (X a) (X (Adbl2 a))

Total

xAddTtl :: Singleton (Root a) => X N -> X a -> XAdd a

random variable for total Additive structures.

Direction

xoAdd :: FibredOriented a => X N -> XOrtOrientation a -> XAdd a

the induced random variable for Additive structures with the given random variable to validate ntimes.

xoRoot :: FibredOriented a => XOrtOrientation a -> X (Root a)

the induced random variables for roots.

xoAdbl2 :: FibredOriented a => XOrtOrientation a -> X (Adbl2 a)

the induced random variable for two addable summands.

xoAdbl3 :: FibredOriented a => XOrtOrientation a -> X (Adbl3 a)

the induced random variable for three addable summands.

xoAbl :: FibredOriented a => X Z -> XOrtOrientation a -> XAbl a

the induced random variable for Abelian structures with the given random variable to validate ztimes.

xoFbr :: XOrtOrientation f -> XFbr f

the associated random variable for validating Fibred structures.

Orientation

xAddOrnt :: Entity p => X N -> X p -> XAdd (Orientation p)

random variable for the Additive structure of Orientation p.

Stalk

xAddStalk :: Additive a => XStalk a -> X N -> XAdd a

random variable for Additive structure.

xStalkAdbl2 :: Additive a => XStalk a -> Root a -> X (Adbl2 a)

random variable of two addable summands rooted at the given root.

xStalkAdbl3 :: Additive a => XStalk a -> Root a -> X (Adbl3 a)

random variable of three addable summands rooted in the given root.