OAlg.Structure.Algebraic.Definition

Description

definition of algebraic structures, i.e. Distributive structures with a suitable Vectorial structure.

Algebraic

class (Distributive a, Vectorial a) => Algebraic a

Distributive structures with a suitable Vectorial structure.

Property Let a be a Algebraic structure, then holds:

For all r in Scalar a and a, b in a with start a == end b holds: r!(a*b) == (r!a)*b and r!(a*b) == a*(r!b).

data Alg k

type representing the class of k-Algebraic structures.

class (ForgetfulOrt (s k), ForgetfulMlt (s k), ForgetfulFbr (s k), ForgetfulFbrOrt (s k), ForgetfulAdd (s k), ForgetfulDst (s k), ForgetfulVec k s, Transformable (s k) (Alg k)) => ForgetfulAlg k s

transformable to k-Algebraic structure.