OAlg.Structure.Distributive.Definition

Description

distributive structures, i.e. multiplicative structures with a suitable additive structure.

Distributive

class (FibredOriented d, Additive d, Multiplicative d) => Distributive d

FibredOriented structures equipped with an Additive and Multiplicative structure satisfying the laws of distributivity.

Properties Let d be a Distributive structure, then holds:

For all g in d and r in Root d with end g == start r holds: zero r * g == zero r' where r' == start g :> end r.

For all g, a and b in d with root a == root b and start a == end g holds: (a + b) * g == a*g + b*g.

For all f in d and r in Root d with start f == end r holds: f * zero r == zero r' where r' = start r :> end f.

For all f, a and b in d with root a == root b and start f == end a holds: f*(a + b) == f*a + f*b.

Note If d is interpreted as a small category C then it is usually called preadditive. If d is also Abelian then C is also usually called abelian.

data Dst

type representing the class of Distributive structures.

class (ForgetfulOrt s, ForgetfulMlt s, ForgetfulFbr s, ForgetfulAdd s, ForgetfulFbrOrt s, Transformable s Dst) => ForgetfulDst s

transformable to Distributive structure.

Transposable

class (TransposableMultiplicative d, Distributive d) => TransposableDistributive d

transposable distributive structures.

Property Let d be a TransposableDistributive structure, then holds:

For all r in Root d holds: transpose (zero r) == zero (transpose r)

For all a, b in d with root a == root b holds: transpose (a + b) == transpose a + transpose b.