OAlg.Structure.Exponential

Description

Multiplicative structures with a power function (^).

Exponential

class (Multiplicative f, Number (Exponent f)) => Exponential f where

Multiplicative structures with a partially defined power function with numbers as exponents.

Properties

For all f and a holds:

If start f == end f or a is an element of [-1,1] then f^a is valid.

If start f /= end f and a is not an element of [-1,1] then f^a is not valid and its evaluation will end up in a NotExponential-exception.

For all f holds: f^1 == f.

For all f holds: f^(-1) == invert f.

For all f and a with start f == end f and a not in [-1,1] holds: start (f^a) == start f and end (f^a) == end f.

For all f, a and b with start f == end f holds: f^(a*b) == (f^a)^ b.

For all f with start f == end f holds: f^0 == one (end f).

For all f, a and b with start f == end f holds: f^(a + b) == f^a * f^b.

For all a and p holds: (one p)^a == one p.

For all f, g and a with start f == end f, start g == end g start f == start g and f * g == g * f holds: (f * g)^a == f^a * g^a.

Note

The phrase ..a is an element of [-1,1].. for the properties of ^ is meant to be: isOne a or isMinusOne a.

If -1 is an instance of Exponent f (see minusOne) then f has to be Cayleyan.

Associated Types

type Exponent f

the exponent.

Methods

(^) :: f -> Exponent f -> f infixl 9

the power of a factor to an exponent.

opower :: (Entity p, Number r) => Orientation p -> r -> Orientation p

the power of an orientation by an number.

Note opower fulfill the properties of Exponential for any number structure.

Real

class Multiplicative f => Real f where

reals.

Methods

power :: Number r => f -> r -> f