OAlg.Structure.Fibred.Definition

Description

fibred structures, i.e. type f with an associated root type Root f such that every value in f has a root.

Fibred

class (Entity f, Entity (Root f)) => Fibred f where

types with a Fibred structure. An entity of a Fibred structure will be called a stalk.

Note

On should accept the default for root only for FibredOriented structures!

For Distributive structures the only thing to be implemented is the Root type and should be defined as Root d = Orientation p where-- p = Point d (see the default implementation of root).

Minimal complete definition

Nothing

Associated Types

type Root f

the type of roots.

Methods

root :: f -> Root f

the root of a stalk in f.

default root :: (Root f ~ Orientation (Point f), Oriented f) => f -> Root f

data Fbr

type representing the class of Fibred structures.

class Transformable s Fbr => ForgetfulFbr s

transformable to Fibred structure.

Fibred Oriented

class (Fibred d, Oriented d, Root d ~ Orientation (Point d)) => FibredOriented d

Fibred and Oriented structure with matching root and orientation.

Property Let d be a FibredOriented structure, then holds: For all s in d holds: root s == orientation s.

Note FibredOriented structures are required for Distributive structures.

data FbrOrt

type representing the class of FibredOriented structures.

class (ForgetfulFbr s, ForgetfulOrt s, Transformable s FbrOrt) => ForgetfulFbrOrt s

transformable to FibredOriented structure.

Spezial classes

class Ord (Root f) => OrdRoot f

type where the associated root type is ordered.

Note Helper class to circumvent undecidable instances.

class Singleton (Root f) => TotalRoot f

type where the associated root type is a singleton.

Sheaf

data Sheaf f

a list in a Fibred structure having all the same root.

Definition Let f be a Fibred structure and s = Sheaf r [t 0 .. t (n-1)] a sheaf in Sheaf f, then s is valid if and only if

r is valid and t i are valid for all i = 0..n-1.

root (t i) == r for all i = 0..n-1.

furthermore n is called the length of s.

If two sheafs have the same root then there stalks can be composed - via (++) - to a new sheaf having the same root. But as (++) is not commutative they are equipped with a Multiplicative structure.

Constructors

Sheaf (Root f) [f]