OAlg.Structure.Fibred.Proposition

Description

propositions on Fibred structures.

Fibred

prpFbr :: Fibred f => XFbr f -> Statement

validity for Fibred structures.

type XFbr = X

random variable for validating Fibred structures.

Fibre Oriented

prpFbrOrt :: FibredOriented f => XFbrOrt f -> Statement

validity for FibredOriented structures.

type XFbrOrt = X

random variable type for validating FibredOriented structures.

X

Orientation

xFbrOrnt :: X p -> XFbr (Orientation p)

random variable for the Fibred structure of Orientation p.

xFbrOrtOrnt :: X p -> XFbrOrt (Orientation p)

random variable for the FibredOriented structure of Orientation p.

xStalkOrnt :: X p -> XStalk (Orientation p)

random variable of XStalk on Orientation p.

Stalk

data XStalk x

random variable for stalks.

Constructors

XStalk (X (Root x)) (Root x -> X x)

xRoot :: XStalk x -> X (Root x)

the underlying random variable of roots.

xSheafRootMax :: Fibred f => XStalk f -> N -> Root f -> X (Sheaf f)

random variable of sheafs in a Fibred structure all rooted in the given root and with a length of either 0 - for empty roots - or with the given length.

xSheafRoot :: Additive a => XStalk a -> N -> Root a -> X (Sheaf a)

random variable of sheafs, all based on the given root and with the given length.

xSheafMax :: Fibred f => XStalk f -> N -> X (Sheaf f)

random variable of sheafs, based on the underlying random variable of roots, with a length of either 0 - for empty roots - or with the given length.

xSheaf :: Additive a => XStalk a -> N -> X (Sheaf a)

random variable of sheafs with the given length.

Fibred Oriented

xoFbr :: XOrtOrientation f -> XFbr f

the associated random variable for validating Fibred structures.

xoFbrOrt :: XOrtOrientation f -> XFbrOrt f

the associated random variable for validation FibredOriented structures.