OAlg.Structure.Multiplicative.Proposition

Description

Propositions on Multiplicative structures.

Multiplicative

prpMlt :: Multiplicative c => XMlt c -> Statement

validity of the Multiplicative structure of c.

data XMlt c

random variable for Multiplicative structures.

Note As the multiplication could by costly, it is recommended to use a bounded random variable for xMltN which serves to validate npower.

Constructors

XMlt

Fields

xMltN :: X N

xMltPoint :: X (Point c)

xMltFactor :: X c

xMltEndo :: X (Endo c)

xMltMltp2 :: X (Mltp2 c)

xMltMltp3 :: X (Mltp3 c)

newtype Endo q

predicate for endos.

Constructors

Endo q

data Mltp2 c

predicate for two multiplicable factors.

Constructors

Mltp2 c c

data Mltp3 c

predicate for three multiplicable factors.

Constructors

Mltp3 c c c

prpMlt1 :: Multiplicative c => p c -> X (Point c) -> Statement

validity of one according to OAlg.Structure.Multiplicative.Definition.

prpMlt2 :: Multiplicative c => X (c, c) -> Statement

validity of * according to OAlg.Structure.Multiplicative.Definition.

prpMlt2_1 :: Multiplicative c => Mltp2 c -> Statement

validity of * for two multiplicable factors according to OAlg.Structure.Multiplicative.Definition.

prpMlt2_2 :: Multiplicative c => c -> c -> Statement

validity of * for two not multiplicable factors according to OAlg.Structure.Multiplicative.Definition.

prpMlt3 :: Multiplicative c => X c -> Statement

validity according to OAlg.Structure.Multiplicative.Definition.

prpMlt4 :: Multiplicative c => X (Mltp3 c) -> Statement

validity according to OAlg.Structure.Multiplicative.Definition.

prpMlt5 :: Multiplicative c => X N -> X c -> X (Endo c) -> Statement

validity of npower according to OAlg.Structure.Multiplicative.Definition.

Note As the multiplication can by very costly the random variable for X N - which serves to check OAlg.Structure.Multiplicative.Definition - has to be chosen carefully.

X

class XStandardMlt c where

standard random variable for Multiplicative structures.

Methods

xStandardMlt :: XMlt c

xMlt :: Multiplicative c => XOrtSite d c -> X N -> X (Endo c) -> XMlt c

random variable for Multiplicative structures.

xMltp2 :: Multiplicative c => XOrtSite d c -> X (Mltp2 c)

random variable of two multiplicable factors.

xMltp3 :: Multiplicative c => XOrtSite d c -> X (Mltp3 c)

random variable of three multiplicable factors.

xoMlt :: Multiplicative c => X N -> XOrtOrientation c -> XMlt c

the induced random variable for multiplicable structures.

Oriented Direction

Total

xMltTtl :: Singleton (Point c) => X N -> X c -> XMlt c

random variable for total Multiplicative structures.

Orientation

xMltOrnt :: Entity p => X N -> X p -> XMlt (Orientation p)

random variable for the Multiplicative structure of Orientation p.