OAlg.Structure.Operational

Description

operations on entities by a Multiplicative structure.

Operation

class Opr f x where

right operation of f on x. This class is rather technical, because on this abstract level it is not possible to define the exact behavior of the operation, i.e. for which values f and x the expression x <* f is valid. For a precise definition see for example TotalOpr or OrientedOpr where the behavior can be stated.

Methods

(<*) :: x -> f -> x infixl 5

right operation.

class Opl f x where

left operation of f on x. This class is rather technical, because on this abstract level it is not possible to define the exact behavior of the operation, i.e. for which values f and x the expression f *> x is valid. For a precise definition see for example TotalOpl or OrientedOpl where the behavior can be stated.

Methods

(*>) :: f -> x -> x infixr 9

Total

class (Opr f x, Multiplicative f, Total f, Entity x) => TotalOpr f x

right operation of a Total Multiplicative structure f on x.

Property Let f be a Total Multiplicative structure and x an instance of Entity, then holds:

For all x in x holds: x <* one u == x where u = unit is the singleton element in Point f.

For all x in x and f, g in f holds: x <* f <* g == x <* (f * g).

Note If f is invertible, then it gives rise of a bijection <* f on x with inverse <* invert f.

class (Opr f x, Multiplicative f, Total f, Entity x) => TotalOpl f x

left operation of a Total Multiplicative structure f on x.

Property Let f be a Total Multiplicative structure and x an instance of Entity, then holds:

For all x in x holds: one u *> x == x where u = unit is the singleton element in Point f.

For all x in x and f, g in f holds: f *> g *> x == (f * g) *> x.

Note If f is invertible, then it gives rise of a bijection f *> on x with inverse invert f *>.

Oriented

class (Opr f x, Multiplicative f, Oriented x) => OrientedOpr f x

right operation of a Multiplicative structure f on a Oriented structure x.

Property Let f be a Multiplicative and x a Oriented structure, then holds:

For all f in f and x in x holds.

If start x == end f then x <* f is valid and orientation (x <* f) == start f :> end x.

If start x /= end f then x <* f is not valid and its evaluation will end up in a NotApplicable exception.

For all x in x holds: x <* one (start x) == x.

For all x in x and f, g in f with end f == start x, end g == start f holds: x <* f <* g == x <* (f * g).

Note If f is invertible, then it rise of a bijection <* f from all x in x with start x == end f to all y in x with start y == start f. Its inverse is given by <* invert f.

class (Opl f x, Multiplicative f, Oriented x) => OrientedOpl f x

left operation of a Multiplicative structure f on a Oriented structure x.

Property Let f be a Multiplicative and x a Oriented structure, and f, x an instance of OrientedOpl, then holds:

For all f in f and x in x holds.

If end x == start f then f *> x is valid and orientation (f *> x) == start x :> end f.

If end x /= start f then f *> x is not valid and its evaluation will end up in a NotApplicable exception.

For all x in x holds: one (end x) *> x== x.

For all x in x and f, g in f with start g == end x, start f == end g holds: f *> g *> x == (f * g) *> x .

Note If f is invertible, then it rise of a bijection f *> from all x in x with end x == start f to all y in x with end y == end f. Its inverse is given by invert f *>.