OAlg.Structure.Ring.Definition

Description

definition of Rings.

Semiring

type Semiring r = (Distributive r, Total r)

distributive structure where * and + are total.

Note If r is Total and Distributive then Root r is Singleton.

rOne :: Semiring r => r

the neutral element according to *, i.e. rOne == one unit.

rZero :: Semiring r => r

the neutral element according to +, i.e. rZero == zero unit.

isMinusOne :: Semiring r => r -> Bool

check being the additive inverse of rOne.

Ring

type Ring r = (Semiring r, Abelian r)

abelian semi rings.

Galoisian

type Galoisian r = (Ring r, Commutative r, Invertible r)

Note Not every element not equal to zero has to be invertible. As such Z is Galoisian.

Field

class Galoisian r => Field r where

not degenerated commutative rings where every element not equal to zero has a multiplicative inverse.

Properties

rZero /= rOne.

For all x and y in r holds:

If y /= rZero then x / y is valid

If y == rZero then x / y is not valid and its evaluation will end up in a NotInvertible-exception.

For all x and y in r with y /= rZero holds: y * (x / y) == x.

Minimal complete definition

Nothing

Methods

(/) :: r -> r -> r infixl 7

division.