OAlg.Structure.Vectorial.Definition

Description

definition of vectorial structures, i.e. Additive structures with a scalar multiplication (!).

Vectorial

class (Semiring (Scalar v), Commutative (Scalar v), Additive v) => Vectorial v where

Additive structures with a total defined scalar multiplication from the left by a commutative semi ring. The entities of v are called vector.

Properties Let v b a Vectorial structure, then holds:

For all s in Scalar v and v in v holds: s!v is valid and root (s!v) == root v.

For all v in v holds: 0!v == zero (root v).

For all s in Scalar v and r in Root v holds s!zero r == zero r.

For all r, s in Scalar v and v in v holds: (r + s)!v == r!v + s!v.

For all s in Scalar v and v, w in v with root v == root w holds: s!(v + w) == s!v + s!w.

For all v in v holds: 1!v == v.

For all r, s in Scalar v and v in v holds: (r*s)!v == r!(s!v).

Associated Types

type Scalar v

the type of scalars.

Methods

(!) :: Scalar v -> v -> v infixr 8

scalar multiplication of a vector.

data Vec k

type representing the class of k-Vectorial structures.

class (ForgetfulFbr (s k), ForgetfulAdd (s k), Transformable (s k) (Vec k)) => ForgetfulVec k s

transformable to k-Vectorial structure.

Euclidean

class Vectorial v => Euclidean v where

Vectorial structures with a partially defined scalar product.

Properties

For all v, w holds: if root v == root w then v <!> w is valid, otherwise a UndefinedScalarproduct-exception will be thrown.

For all u holds: u <!> zero (root u) == rZero.

For all u, v and w with root u == root w and root w == root v holds: u <!> (v + w) == u <!> v + u <!> w.

For all w holds: zero (root w) <!> w == rZero.

For all u, v and w with root w == root u and root u == root v holds: (u + v) <!> w == u <!> w + v' !' w.

Methods

(<!>) :: v -> v -> Scalar v infix 7

the scalar product of two vectors.