OAlg.Category.Definition

Description

categories of morphisms. We adapted the concept of categories form Category to better cover our needs.

Category

class Morphism c => Category c where

category of morphisms.

Properties Let c be a type instance of the class Category, then holds:

For all types x, y and f in c x y holds: cOne (range f) . f = f and f . cOne (domain f) = f.

For all types w, x, y, z and f in c x w, g in c y x, h in c z y holds: f . (g . h) = (f . g) . h.

Methods

cOne :: Struct (ObjectClass c) x -> c x x

the identity morphism for an eligible x.

(.) :: c y z -> c x y -> c x z infixr 9

cOne' :: Category c => p c -> Struct (ObjectClass c) x -> c x x

the cOne to a given Struct (ObjectClass c). The type p c serves only as proxy and cOne' is lazy in it.

Note As ObjectClass may be a non-injective type family, the type checker needs some times a little bit more information to pic the right cOne.

Some basic definitions in the category (->)

id :: x -> x

the identity map.

const :: b -> a -> b

the constant map given by a value in b.

Property Let y be in b then for all x in a holds: const y x is identical to y.

curry :: ((a, b) -> c) -> a -> b -> c

currying a map.

uncurry :: (a -> b -> c) -> (a, b) -> c

uncurrying a map.

fst :: (a, b) -> a

the first component of the pair.

snd :: (a, b) -> b

the second component of the pair.

curry3 :: ((a, b, c) -> d) -> a -> b -> c -> d

currying a map.

uncurry3 :: (a -> b -> c -> d) -> (a, b, c) -> d

uncurrying a map.

Cayleyan

class (Category c, Eq2 c) => Cayleyan2 c where

category of isomorphisms.

Property Let c be a type instance of Cayleyan2, then holds: For all types x, y and f in c x y holds: (invert2 f . f) == cOne (domain f) and (f . invert2 f) == cOne (range f) where (==) = eq2.

Methods

invert2 :: c x y -> c y x

Morphism

class Morphism m where

morphism.

Minimal complete definition

homomorphous | domain, range

Associated Types

type ObjectClass m

the object class.

Methods

homomorphous :: m x y -> Homomorphous (ObjectClass m) x y

attests, that the types x and y fulfill the constraints given by Homomorphous (ObjectClass m) x y, i.e both fulfill the constraints given by Structure (ObjectClass m) x and Structure (ObjectClass m) y respectively.

domain :: m x y -> Struct (ObjectClass m) x

attests that the domain type x fulfills the constraints given by Structure (ObjectClass m) x.

range :: m x y -> Struct (ObjectClass m) y

attests that the range type y fulfills the constraints given by Structure (ObjectClass m) y.

data Homomorphous s x y

attest that both x and y have homomorphous structures, i.e. both admit the same constraints given by the parameter s.

Constructors

(Struct s x) :>: (Struct s y) infix 5

tauHom :: Transformable s t => Homomorphous s x y -> Homomorphous t x y

transforming homomorphous structural attests.

tau1Hom :: Transformable1 f s => Homomorphous s x y -> Homomorphous s (f x) (f y)

transforming homomorphous structural attests.

eqlDomain :: Struct Typ x -> Struct Typ x' -> m x y -> m x' y -> Maybe (x :~: x')

gets for two Typeable types x and x' and for two parameterized types maybe an attest that the domain types are equal.

eqlRange :: Struct Typ y -> Struct Typ y' -> m x y -> m x y' -> Maybe (y :~: y')

gets for two Typeable types y and y' and for two parameterized types maybe an attest that the range types are equal.

eqlMorphism :: Typeable m => Struct Typ x -> Struct Typ x' -> Struct Typ y -> Struct Typ y' -> m x y -> m x' y' -> Maybe (m x y :~: m x' y')

gets maybe an attest that the two given morphisms types are equal.

Applicative

class Applicative h where

family of types having a representation in (->).

Methods

amap :: h a b -> a -> b

application.

($) :: Applicative h => h a b -> a -> b infixr 0

right associative application on values.

class Applicative1 h f where

family of types having a representation in f a -> f b.

Methods

amap1 :: h a b -> f a -> f b

application.

Functorial

class (Applicative c, Category c) => Functorial c

representable categories, i.e. covariant functors from an Applicative category c to (->).

Properties Let c be a type instance of the class Functorial then holds:

For all types x and d in Struct (ObjectClass c) x holds: amap (cOne d) = id.

For all types x, y, z and f in c y z, g in c x y holds: amap (f . g) = amap f . amap g.

Forget

data Forget t m x y where

forgets the ObjectClass of m and sets it to t, under the condition that the ObjectClass of m is Transformable to t.

Constructors

Forget :: Transformable (ObjectClass m) t => m x y -> Forget t m x y

Embeddable

class (Morphism m, Transformable (ObjectClass m) t) => EmbeddableMorphism m t

morphism for which its object class can be embedded into the given structure.

class EmbeddableMorphism m Typ => EmbeddableMorphismTyp m

helper class to avoid undecidable instances.