OAlg.Category.Unify

Description

unification of categories, i.e. projecting morphisms by dropping the parameterization by there domain and range.

Morphism

data SomeMorphism m where

some morphism.

Constructors

SomeMorphism :: m x y -> SomeMorphism m

data SomeObjectClass m where

some object class.

Constructors

SomeObjectClass :: Transformable (ObjectClass m) Typ => Struct (ObjectClass m) x -> SomeObjectClass m

data SomeMorphismSite s m x where

some morphism given by a Site.

Constructors

SomeMorphismDomain :: m x y -> SomeMorphismSite From m x

SomeMorphismRange :: m x y -> SomeMorphismSite To m y

Path

data SomePath m where

some path

Constructors

SomePath :: Path m x y -> SomePath m

somePath :: SomePathSite s m x -> SomePath m

embedding.

data SomePathSite s m x where

some path parameterized either by its domain or range.

Constructors

SomePathDomain :: Path m x y -> SomePathSite From m x

SomePathRange :: Path m x y -> SomePathSite To m y

Entity

data SomeEntity m where

some entity x in x having the given ObjectClass m as structure.

Constructors

SomeEntity :: Entity x => Struct (ObjectClass m) x -> x -> SomeEntity m

Application

data SomeApplication h where

some application.

Constructors

SomeApplication :: h x y -> x -> SomeApplication h