OAlg.Data.Constructable

Description

constructing values by there Form..

Constructable

class Exposable x => Constructable x where

types with an associated form, which serves as a blueprint to construct a corresponding value.

A common setting for this structure is a module with a reducible type f (see Reducible) with public constructors - which serves as a form to be filled out - and in the same module a type e with a private constructor - lets say E - to hold the reduced f. Than an implementation would be

make f = E (reduce f)

and

form (E f) = f

Property Let x be an instance of the class Constructable than holds: For all x in x holds: make (form x) == x.

Methods

make :: Form x -> x

constructor.

cmap :: (Constructable x, Constructable y) => (Form x -> Form y) -> x -> y

restriction of a f in Form x -> Form y.

Exposable

class Exposable x where

types with an associated form type and a function form which exposes its values to its form.

Associated Types

type Form x

the form.

Methods

form :: x -> Form x

the associated form.

restrict :: Exposable x => (Form x -> y) -> x -> y

restriction of a f in Form x -> y.