OAlg.Data.Dualisable

Description

data admitting a kind of duality.

Dual

type family Dual (x :: k) :: k

the assigned dual kind.

Dualisable

class Dualisable x where

admitting a duality.

Property Let x be Dualisable, than holds: toDual is a bijection with its inverse fromDual.

Methods

toDual :: x -> Dual x

fromDual :: Dual x -> x

fromDual' :: Dualisable x => p x -> Dual x -> x

fromDual enriched with a parameterized type p which serves as a proxy - e.g. Proxy or Id will serve - and will not be evaluated. It serves for the type checker to pick the right fromDual.

Reflexive

class Reflexive x where

admitting reflection.

Property Let x be Reflexive, than holds:

toBidual is a bijection with its inverse fromBidual.

Methods

toBidual :: x -> Dual (Dual x)

fromBidual :: Dual (Dual x) -> x

fromBidual' :: Reflexive x => p x -> Dual (Dual x) -> x

fromBidual enriched with a parameterized type p which serves as a proxy - e.g. Proxy or Id will serve - and will not be evaluated. It serves for the type checker to pick the right fromBidual.

Transposable

class Transposable x where

transposable types..

Property Let x be a Transposable, then holds: For all x in x holds: transpose (transpose x) == x.

Methods

transpose :: x -> x

Site

data Site

concept of the sites From and To.

Constructors

From

To

Side

data Side

concept of sides LeftSide and RightSide

Constructors

LeftSide

RightSide

Direction

data Direction

concept of the directions LeftToRight and RightToLeft.

Constructors

LeftToRight

RightToLeft