OAlg.Data.Ord

Description

Data.Ord enriched with some additional elements.

Total

fcompare :: Ord i => (a -> i) -> a -> a -> Ordering

comparing according to the mapped values.

wcompare :: (w -> w -> Ordering) -> (a -> w) -> a -> a -> Ordering

comparing according to the given ordering relation on the mapped values.

coCompare :: (a -> a -> Ordering) -> a -> a -> Ordering

the reverse ordering

compare2 :: (a -> a -> Ordering) -> (b -> b -> Ordering) -> (a, b) -> (a, b) -> Ordering

comparing of pairs.

sortFst :: Ord a => [(a, b)] -> [(a, b)]

sorting according to the first component.

sortFstBy :: (a -> a -> Ordering) -> [(a, b)] -> [(a, b)]

sorting according to the first component.

sortSnd :: Ord b => [(a, b)] -> [(a, b)]

sorting according to the second component.

sortSndBy :: (b -> b -> Ordering) -> [(a, b)] -> [(a, b)]

sorting according to the second component.

data Closer x

the closer of a linear ordered x.

Constructors

NegInf

It x

PosInf

cmin :: Ord x => [x] -> Closer x

the minimum of the items of a list, i.e. the biggest lower bound.

Property Let xs be in [x] for a linear ordered x, then holds: For all x in xs holds: cmin xs <= It x.

cmax :: Ord x => [x] -> Closer x

the maximum of the items of a list, i.e. the smallest upper bound.

Property Let xs be in [x] for a linear ordered x, then holds: For all x in xs holds: It x <= cmax xs.

cspan :: Ord x => [x] -> Span x

(l,u) = cspan xs where l is the minimum and u the maximum of the items in xs.

Example

>>> cspan "aeb"
(It 'a',It 'e')

>>> cspan ""
(PosInf,NegInf)

type Span x = (Closer x, Closer x)

the span type.

enumSpan :: (Enum i, Ord i) => i -> Closer i -> [i]

enumSpan i0 h enumerates the index, starting by i0 to h.

Partial

class Eq a => POrd a where

partially ordered types.

Properties Let a be an instance of POrd, then holds:

For all x in a holds: x <<= x.

For all x, y in a holds: If x <<= y and y <<= x then x == y.

For all x, y, z in a holds: If x <<= y and y <<= z then x <<= z.

Methods

(<<=) :: a -> a -> Bool infix 4