OAlg.Entity.Diagram.Transformation

Description

natural transformations between Diagrams.

Transformation

data Transformation t n m a

natural transformations between two Diagrams.

Property Let Transformation a b t be in Transformation t n m a for a Multiplicative structure a, then holds

dgQuiver a == dgQuiver b.

For all 0 <= i < n holds: orientation (t i) == p i :> q i where p = dgPoints a and q = dgPoints b.

For all 0 <= j < m holds: t (e j) * f j == g j * t (s j) where f = dgArrows a, g = dgArrows b, s j is the index of the start point of the j-th arrow and e j is the index of the end point.

                   t (s j)
   s j     p (s j) --------> q (s j)
    |         |                 |
  j |     f j |                 | g j
    |         |                 |
    v         v                 v
   e j     p (e j) --------> q (e j)
                   t (e j)

Constructors

Transformation (Diagram t n m a) (Diagram t n m a) (FinList n a)

trfs :: Transformation t n m a -> FinList n a

the underlying list of factors.

Duality

coTransformation :: Transformation t n m a -> Dual (Transformation t n m a)

the dual transformation.