OAlg.Entity.FinList

Description

finite lists, parameterized by there length.

FinList

data FinList (n :: N') a where

finite lists, parameterized by there length.

Constructors

Nil :: FinList N0 a

(:|) :: a -> FinList n a -> FinList (n + 1) a infixr 5

toW :: FinList n a -> W n

the underlying witness.

head :: FinList (n + 1) a -> a

the head of a non empty finite list.

tail :: FinList (n + 1) a -> FinList n a

the tail of a non empty finite list.

zip :: FinList n a -> FinList n b -> FinList n (a, b)

zips two sequences of the same length.

zip3 :: FinList n a -> FinList n b -> FinList n c -> FinList n (a, b, c)

zips three sequences of the same length.

(|:) :: FinList n a -> a -> FinList (n + 1) a infixl 5

appending an element at the end of the finite list.

(++) :: FinList n a -> FinList m a -> FinList (n + m) a infixr 5

appending two finite lists.

(**) :: FinList n a -> FinList m b -> FinList (n * m) (a, b)

the product of two finite lists.

repeat :: Any n -> a -> FinList n a

the constant sequence.

iFinList :: N -> FinList n a -> FinList n (a, N)

indexing finite lists, starting at the given natural number.

iFinList0 :: FinList n a -> FinList n (a, N)

adjoins to each element its index, starting from '0'.

iFinList' :: Any n -> FinList n N

the sequence 0,1 .. n-1.

toArray :: FinList n a -> Array N a

maps a sequnece as = a0..a(n-1) of length n to the corresponding array a with ai == a$i for i = 0..(n-1).

someFinList :: [a] -> SomeFinList a

the associated finite list.

data SomeFinList a

some finite list.

Constructors

forall n. SomeFinList (FinList n a)

(<++>) :: SomeFinList x -> SomeFinList x -> SomeFinList x infixr 5

concatenation.

Induction

inductionS :: Any n -> FinList N0 a -> (forall i. FinList i a -> FinList (i + 1) a) -> FinList n a

induction for sequences.

newtype FinList' a n

Wrapper to switch the parameters.

Constructors

FinList' (FinList n a)