OAlg.Entity.Matrix.Definition

Description

defintion of matrices over Distributive structures.

Matrix

data Matrix x

matrix over Distributive structures.

Property Let Matrix rw cl xijs be in Matrix a for a Distributive structure a, then holds:

rw and cl are valid.

xijs is valid.

For all (x,i,j) in xijs holds:

x is not zero.

orientation x == (cl ? j) :> (rw ? i).

Constructors

Matrix (Dim' x) (Dim' x) (Entries N N x)

rows :: Matrix x -> Dim' x

row dimension.

cols :: Matrix x -> Dim' x

column dimension.

mtxxs :: Matrix x -> Entries N N x

the entries.

mtxRowCol :: Matrix x -> Row N (Col N x)

viewing as a row of columns.

mtxColRow :: Matrix x -> Col N (Row N x)

viewing as a column of rows.

mtxMap :: Hom Dst h => h x y -> Matrix x -> Matrix y

mapping of a matrix.

Group

mtxGroupRow :: Distributive x => Matrix x -> Matrix (Matrix x)

groups the rows with same row dimensions into a matrix of matrices with one column and n rows accordingly.

mtxGroupDim :: Distributive x => Dim' x -> Dim' (Matrix x)

groups a formal product of points p 0 ^ r 0 * .. * p n ^ r n into a formal product of dimensions (dim [p 0] ^ r o) * .. * (dim [p n] ^ r n).

mtxJoin :: Oriented x => Matrix (Matrix x) -> Matrix x

joining block matrices, i.e. matrices of matrices.

mtxJoinDim :: Oriented x => Dim' (Matrix x) -> Dim' x

joining the dimension of matrices over x.

Construction

matrix :: (Additive x, p ~ Point x) => Dim x p -> Dim x p -> [(x, N, N)] -> Matrix x

matrix with the given row and column number and the given entries for a Distributive structure.

Property Let m = matrix rw cl xis then holds

Pre

For all (x,i,j) in xijs holds: start x == cl ? j and end x == rw ? i.

Post

m is valid.

Note The given entries will be sorted, aggregated and zeros eliminated.

matrixTtl :: (Additive x, FibredOriented x, Total x) => N -> N -> [(x, N, N)] -> Matrix x

matrix with the given row and column number and the given entries for a Total Distributive structure.

Property Let m = matrixTtl rws cls xis then holds

Pre

For all (_,i,j) in xijs holds: i < rws and j < cls.

Post

m is valid.

Note The given entries will be sorted, aggregated and zeros eliminated.

matrixBlc :: (Additive x, FibredOriented x) => [Dim' x] -> [Dim' x] -> [(Matrix x, N, N)] -> Matrix (Matrix x)

block matrices as matrix of matrices.

diagonal :: Additive x => Dim' x -> Dim' x -> [x] -> Matrix x

diagonal matrix with entries starting at the index 0 (see diagonal').

diagonal' :: Additive x => N -> Dim' x -> Dim' x -> [x] -> Matrix x

diagonal matrix with entries starting at the given index offset.

Duality

coMatrix :: Entity (Point x) => Matrix x -> Dual (Matrix x)

the dual matrix, with inverse coMatrixInv.

coMatrixInv :: Entity (Point x) => Dual (Matrix x) -> Matrix x

from the dual matrix, with inverse coMatrix.

mtxFromOpOp :: Entity (Point x) => Matrix (Op (Op x)) -> Matrix x

from the bidual.

Homomorphisms

isoCoMatrixDst :: Distributive x => IsoOpMap Matrix Dst (Op (Matrix x)) (Matrix (Op x))

the contravariant isomorphism from Matrix x to Matrix (Op x).

X

class XStandardOrientationMatrix x where

standard random variable for the orientations of matrices over x.

Methods

xStandardOrientationMatrix :: X (Orientation (Dim' x))

xMatrix :: Additive x => Q -> XOrtOrientation x -> X (Orientation (Point (Matrix x))) -> XOrtOrientation (Matrix x)

random variable of matrices with the given maximal dimension and density.

xMatrixTtl :: (Distributive x, Total x) => N -> Q -> X x -> XOrtOrientation (Matrix x)

random variable of matrices with the given maximal dimension and the given density.

Direction

xodZ :: XOrtOrientation (Matrix Z)

a random variable of Z-matrices.

xodZZ :: XOrtOrientation (Matrix (Matrix Z))

a random variable of Z-bolck-matrices.