OAlg.Entity.Matrix.Dim

Description

dimension for matrices of x as a complete sequence of Point x.

Documentation

data Dim x p where

dimension of x as a complete sequence of Point x.

Constructors

Dim :: CSequence (Point x) -> Dim x (Point x)

type Dim' x = Dim x (Point x)

abbreviation for Dim x (Point x).

fromDim :: Dim x p -> ProductSymbol p

the underlying product.

dim :: (Entity p, p ~ Point x) => p -> Dim x p

constructing a dimension form a point.

productDim :: (Entity p, p ~ Point x) => [p] -> Dim x p

constructing a dimension from a list of points.

dimxs :: p ~ Point x => Dim x p -> [(p, N)]

the indexed listing of the points.

dimwrd :: (Entity p, p ~ Point x) => Dim x p -> Word N p

the underlying word.

dimMap :: (Entity q, q ~ Point y) => (p -> q) -> Dim x p -> Dim y q

mapping a dimension.