OAlg.Entity.Matrix.GeneralLinearGroup

Description

general linear group GL and elementary transformations over a Galoisian structure.

Transformation

data Transformation x where

elementary linear transformation over a Distributive structure x.

Property Let f be in Transformation x then holds:

If f matches Permute r c p then holds:

h <= It n where (_,h) = span nProxy p and n = lengthN c.

r == c <* p.

If f matches Scale d k s then holds:

k < lengthN d.

s is an endo at d ? k.

s is valid.

If f matches Shear d k l g then holds:

k < lengthN d and l < lengthN d.

k < l.

g is valid.

Note Shear d k l (GL2 s t u v) represents the square matrix m of dimension d where m k k == s, m k l == t, m l k == u, m l l == v and for all i, j not in [k,l] holds: If i /= j then m i j is zero else m i i is one.

Constructors

Permute :: Distributive x => Dim x (Point x) -> Dim x (Point x) -> Permutation N -> Transformation x

Scale :: Distributive x => Dim x (Point x) -> N -> Inv x -> Transformation x

Shear :: Galoisian x => Dim x (Point x) -> N -> N -> GL2 x -> Transformation x

GL

type GL x = Inv (Matrix x)

general linear groupoid of matrices.

data GL2 x

the general linear group of 2x2 matrices for a Galoisian structure x.

Property Let GL2 s t u v be in GL2 x for a Galoisian structure x, then holds: s*v - u*t is invertible.

Example Let g = GL2 3 5 4 7 :: GL2 Z:

>>> invert g
 GL2 7 -5 -4 3

>>> g * invert g
GL2 1 0 0 1

which is the one in GL2 Z.

Note

GL2 (s t u v) represents the 2x2-matrix

   [s t]
   [u v]

and is obtained by GL2GL.

Constructors

GL2 x x x x

GLT

data GLT x

quotient groupoid of the free groupoid of Transformation (see FTGLT) given by the relations:

permuteFT d c p * permuteFT b a q ~ permuteFT d a (q*p) where b == c and Permute d c p, Permute b a q are valid (Note: the permutations p and q are switched on the right side of the equation).

...

Property Let g be in GLT, then holds:

For all exponents z in form g holds: 0 < z.

Example Let d = dim [()] ^ 10 :: Dim' Z, a = permuteFT d d (swap 2 8), b = permuteFT d d (swap 2 3) and c = permuteFT d d (swap 2 3 * swap 2 8) then:

>>> a * b == c
False

but in GLT holds: let a' = amap FTGLT a, b' = amap FTGLT b and c' = amap FTGLT c in

>>> a' * b' == c'
True

and

>>> amap GLTGL (a' * b') == amap GLTGL a' * amap GLTGL b'
True

Note: As a consequence of the property (1.), GLT can be canonically embedded via prj . form - in to ProductForm N (Transformation x).

permute :: Distributive x => Dim' x -> Dim' x -> Permutation N -> GLT x

permutation of the given dimensions.

Property Let r, c be in Dim' x and p in Permutation N for a Distributive structure x, then holds: If Permute r c p is valid then permute r c p is valid.

Example Let t = permute r c p with Permute r c p is valid then its associated matrix (see GLTGL) has the orientation c :> r and the form

           k         l
  [1                          ]
  [  .                        ]
  [    .                      ]
  [     1                     ]
  [                 1         ] k
  [         1                 ]
  [           .               ]
  [             .             ]
  [               1           ]
  [       1                   ] l
  [                    1      ]
  [                      .    ]
  [                        .  ]
  [                          1]

Note r dose not have to be equal to c, but from r == c <* p follows that both have the same length.

permuteFT :: Distributive x => Dim' x -> Dim' x -> Permutation N -> FT x

the induce element in the free groupoid of transformations.

scale :: Distributive x => Dim' x -> N -> Inv x -> GLT x

scaling.

Property Let d be in Dim' x, k in N and s in Inv x, then holds: If Scale d k s is valid then scale d k s is valid.

Example Let t = scale d k s with Scale d k s is valid then its associated matrix (see GLTGL) is an endo with dimension d and has the form

          k         
  [1               ]
  [  .             ]
  [    .           ]
  [     1          ]
  [      s'        ] k
  [         1      ]
  [           .    ]
  [             .  ]
  [               1]

where s' = (inj :: Inv x -> x) s.

shear :: Galoisian x => Dim' x -> N -> N -> GL2 x -> GLT x

shearing.

Property Let d be in Dim' x, k, l in N and g in GL2 x then holds: If Shear d k l g is valid then shear d k l g is valid.

Example Let t = shear d k l g where Shear d k l g is valid then its associated matrix (see GLTGL) is an endo with dimension d and has the form

           k         l
  [1                          ]
  [  .                        ]
  [    .                      ]
  [     1                     ]
  [       s         t         ] k
  [         1                 ]
  [           .               ]
  [             .             ]
  [               1           ]
  [       u         v         ] l
  [                    1      ]
  [                      .    ]
  [                        .  ]
  [                          1]

rdcGLTForm :: Oriented x => GLTForm x -> GLTForm x

reduces a GLTForm x to its normal form.

Property Let f be in GLTForm x for a Oriented structure x, then holds:

rdcGLTForm (rdcGLTForm f) == rdcGLTForm f.

For all exponents z in rdcGLTForm f holds: 0 < z.

type GLTForm x = ProductForm Z (Transformation x)

form of GLT.

gltfTrsp :: TransposableDistributive r => GLTForm r -> GLTForm r

transposition of a product of elementary transformation.

FT

type FT x = Product Z (Transformation x)

the free groupoid of Transformations.

Homomorphism

Ort

data TrApp x y where

Oriented homomorphisms.

Constructors

TrFT :: Oriented x => TrApp (Transformation x) (FT x)

TrGL :: Distributive x => TrApp (Transformation x) (GL x)

TrGLT :: Oriented x => TrApp (Transformation x) (GLT x)

trGLT :: Oriented x => Transformation x -> GLT x

the induced element of the groupoid GLT.

Mlt

data GLApp x y where

Multiplicative homomorphisms.

Constructors

FTGL :: Distributive x => GLApp (FT x) (GL x)

FTGLT :: Oriented x => GLApp (FT x) (GLT x)

GLTGL :: Distributive x => GLApp (GLT x) (GL x)

GL2GL :: Galoisian x => GLApp (GL2 x) (GL x)