OAlg.Entity.Matrix.Vector

Description

Vectors with coefficients, lying in a Semiring.

Vector

newtype Vector r

vector with coefficients lying in a Semiring, indexd by N.

Definition Let v = Vector ris be in Vector r with r be a Semiring, then v is valid iff

ris is valid

For all (r,i) in ris holds: r is not equal to rZero.

Constructors

Vector (PSequence N r)

vecpsq :: Vector r -> PSequence N r

the underlying partial sequence.

cf :: Semiring r => Vector r -> N -> r

the i-th coefficient of the given vector.

Example Let v = vector [(-3,2),(9,4)] :: Vector Z

>>> map (cf v) [0..8]
[0,0,-3,0,9,0,0,0,0]

cfsssy :: (Semiring r, Commutative r, Entity a, Ord a) => Set a -> Vector r -> SumSymbol r a

the associated free sum of symbols according to the given set of symbols and coefficients.

Property Let s = s 0 < s 1 < .. be in Set a and r be in Vector r then holds: cfsssy s r == cf r 0 ! sy (s 0) + cf r 1 ! sy (s 1) + ...

ssycfs :: (Semiring r, Ord a) => Set a -> SumSymbol r a -> Vector r

the associated coefficients of a free sum of symbols according to the given set of symbols.

Property Let s = s 0 < s 1 < .. be in Set a and x in SumSymbol r a then holds: ssyprj s x == cf r 0 ! sy (s 0) + cf r 1 ! sy (s 1) + .. where r = ssycfs s x,

vecrc :: Vector r -> Row N (Col N r)

a vector as a row with one column at 0.

vecAppl :: Semiring r => Matrix r -> Vector r -> Vector r

applying a matrix from the left.

Hom

data HomSymbol r x y where

Constructors

HomSymbol :: (Entity x, Ord x, Entity y, Ord y) => PSequence x (LinearCombination r y) -> HomSymbol r (SumSymbol r x) (SumSymbol r y)

Cfs :: (Entity x, Ord x) => Set x -> HomSymbol r (SumSymbol r x) (Vector r)

Ssy :: (Entity x, Ord x) => Set x -> HomSymbol r (Vector r) (SumSymbol r x)

HomMatrix :: Matrix r -> HomSymbol r (Vector r) (Vector r)

mtxHomSymbol :: Matrix r -> HomSymbol r (SumSymbol r N) (SumSymbol r N)

the associated r-linear homomorphism.

Representation

repMatrix :: Representable r h x y -> Matrix r

the associated representation matrix of the given r-homomorphism and the two symbol set.

Property Let p = Representable h xs ys be in Representable r h x y for a Commutative Semiring r, then holds: For all v in Vector r holds: Let h' = HomMatrix (repMatrix p) in

For all (_,i) in h' $ v holds: i < lengthN ys.

(Ssy ys $ h' $ v) == (h $ Ssy xs $ v).

data Representable r h x y where

Predicate for a r-linear homomorphisms between the free sums SumSymbol r x and SumSymbol r y being representable for the given symbol sets.

Definition Let l be in LinearCombination r x and xs be a Set of symbols of x, then l is called representable in xs iff all symbols of lcs l are elements of xs.

Property Let h be a r-linear homomorphism between the free sums SumSymbol r x and SumSymbol r y, xs a Set of symbols in x and ys a Set of symbols in y, then holds: If for each symbol x in xs the associated LinearCombination of h $ x is representable in ys, then Representable h xs ys is valid.

Constructors

Representable :: (Hom (Vec r) h, Entity x, Ord x, Entity y, Ord y) => h (SumSymbol r x) (SumSymbol r y) -> Set x -> Set y -> Representable r h (SumSymbol r x) (SumSymbol r y)

mtxRepresentable :: (Semiring r, Commutative r) => Matrix r -> Representable r (HomSymbol r) (SumSymbol r N) (SumSymbol r N)

the associated representation of a matrix.

Propostion

prpRepMatrix :: (Semiring r, Commutative r) => Representable r h x y -> Vector r -> Statement

validity of repMatrix for the given vector.

prpRepMatrixZ :: N -> N -> Statement

validity of repMatrix for Z-matrices with the given row and column numbers.

X

xVecN :: Semiring r => N -> X r -> X (Vector r)

random variable of Vector r where all indices are strict smaller then the given bound.

Property Let n be in N and xr be in X r then holds: For all (_,i) in the range of xVecN n xr holds: i < n.