OAlg.Entity.Natural

Description

natural numbers promotable to type-level. They serve to parameterize the length of finite lists on type-level (see OAlg.Entity.FinList).

Note We need the language extension UndecidableInstances for type checking the definition

N0 * m = N0
S n * m = m + n * m

for the type family *.

Using UndecidableInstances could cause the type checker to not terminate, but this will obviously not happen for the given definition (we also use the same definition for the multiplicative structure of N').

Natrual Numbers

data N'

natural number promotable to type-level.

Constructors

N0

S N'

toN' :: N -> N'

mapping a natural number in N to N'.

type family (n :: N') + (m :: k) :: N' infixr 6

addition of natural numbers.

type family (n :: k) * (m :: N') :: N' infixr 7

multiplication of natural numbers.

Attest

class Typeable n => Attestable n where

attest for any natural number.

Methods

attest :: W n

data Ats n where

witness of being attestable.

Constructors

Ats :: Attestable n => Ats n

nValue :: Attestable n => p n -> N

the corresponding natural number in N of a type parameterized by n.

data W n where

witness for a natural number and serves for inductively definable elements (see induction).

Constructors

W0 :: W N0

SW :: W n -> W (n + 1)

cmpW :: W n -> W m -> Ordering

comparing of two witnesses.

(++) :: W n -> W m -> W (n + m) infixr 6

addition of witnesses.

(**) :: W n -> W m -> W (n * m) infixr 7

multiplication of witnesses.

data W'

union of all witnesses, which is isomorphic to N and N'.

Constructors

forall n. W' (W n)

toW' :: N -> W'

mapping a natural value in N to W'.

data SomeNatural where

witnesse of some natural.

Constructors

SomeNatural :: Attestable n => W n -> SomeNatural

succSomeNatural :: SomeNatural -> SomeNatural

successor of some natural number.

someNatural :: N -> SomeNatural

mapping N to SomeNatural.

Note The implementation of this mapping is quite inefficent for high values of N.

naturals :: [SomeNatural]

the infinite list of some naturals.

Induction

induction :: Any n -> f N0 -> (forall i. f i -> f (i + 1)) -> f n

induction for general type functions.

Propositions

The following propositions are based - for the most part - on induction. To ensure that the proofs yield a terminal value, i.e. Refl, we used in the proof of the induction hypothesis only propositions which are proofed before. So: if these propositions terminate, then also the hypothesis will terminate.

type Any = W

predicate for any natural number.

type Some = W'

some witness.

Some basic Lemmas

refl :: Any x -> x :~: x

reflexivity.

lemma1 :: (x :~: y) -> f x :~: f y

well definition of parameterized types.

sbstAdd :: (a :~: a') -> (b :~: b') -> (a + b) :~: (a' + b')

substitution rule for addition.

Some abbreviations

prpAdd0 :: (a + 0) :~: a

0 is right neural for the addition.

prpAdd1 :: (a + 1) :~: S a

adding 1 is equal to the successor.

prpAdd2 :: (a + 2) :~: S (S a)

adding 2 is equal to the successor of the successor.

Equivalence of Any

prpEqlAny :: (Any n :~: Any m) -> n :~: m

equality of the underlying natural number.

prpEqlAny' :: (n :~: m) -> Any n :~: Any m

Any is as parameterized type is well defined.

Injectivity of the Successor

prpSuccInjective :: (S n :~: S m) -> n :~: m

successor is injective.

Addition

Neutral Element

N0 is the neutral element of +.

prpAddNtrlL :: p a -> (N0 + a) :~: a

N0 is left neutral for the addition.

prpAddNtrlR :: Any a -> (a + N0) :~: a

N0 is the right neutral for the addition.

Associativity

prpAddAssoc :: Any a -> Any b -> Any c -> ((a + b) + c) :~: (a + (b + c))

addition is associative.

Commutativity

lemmaAdd1 :: Any a -> Any b -> (S a + b) :~: (a + S b)

lemma 1 for the addition.

prpAddComm :: Any a -> Any b -> (a + b) :~: (b + a)

addition is commutative.

Multiplication

Neutral Element

N1 = S N0 is the neutral element of *.

prpMltNtrlL :: Any a -> (N1 * a) :~: a

N1 is left neutral for the multiplication.

prpMltNtrlR :: Any a -> (a * N1) :~: a

N1 is right neutral for the multiplication.

Distributivity on the left

prpDstrL :: Any a -> Any b -> Any c -> ((a + b) * c) :~: ((a * c) + (b * c))

law of left distributivity holds.

Associativity

prpMltAssoc :: Any a -> Any b -> Any c -> ((a * b) * c) :~: (a * (b * c))

multiplication is associative.

Distributivity on the right

lemmaMlt1 :: Any a -> (a * N0) :~: N0

right multiplication of N0 is N0.

lemmaAdd2 :: Any a -> Any b -> Any c -> (a + (b + c)) :~: (b + (a + c))

lemma 2 for addition.

prpDstrR :: Any a -> Any b -> Any c -> (a * (b + c)) :~: ((a * b) + (a * c))

law of right distributivity holds.

Commutativity

prpMltComm :: Any a -> Any b -> (a * b) :~: (b * a)

multiplication is commutative.

Some attests

generating the Haskell code and interface for witnesses.

codeW :: N -> N -> String

codeW n m generates the haskell code for the witnesses W of N' from n to m.

itfW :: N -> N -> String

itfW n m generates the haskell interface for the witnesses W from n to m.

type N0 = 'N0

0.

type N1 = S N0

1.

type N2 = S N1

2.

type N3 = S N2

3.

type N4 = S N3

4.

type N5 = S N4

5.

type N6 = S N5

6.

type N7 = S N6

7.

type N8 = S N7

8.

type N9 = S N8

9.

type N10 = S N9

10.

X

xSomeNatural :: X N -> X SomeNatural

the induced random variable for some natural number.