OAlg.Entity.Product.ProductSymbol

Description

free products of symbols in x with index type N.

ProductSymbol

newtype ProductSymbol x

free product of symbols in x with index type N.

Example

The expression sy 'a' constructs a free product of exactly one symbol in Char consisting just of the character 'a'.

>>> sy 'a'
ProductSymbol['a']

they are Total Multiplicative

>>> sy 'a' * sy 'b' * sy 'c'
ProductSymbol['a'*'b'*'c']

and admit a listing

>>> list (Proxy :: Proxy N) (sy 'a' * sy 'b' * sy 'c')
[('a',0),('b',1),('c',2)]

they have a compact representation for repetitions

>>> sy 'a' * sy 'b' * sy 'b' * sy 'a' * sy 'c'
ProductSymbol['a'*'b'^2*'a'*'c']

>>> sy 'a' * sy 'b' * sy 'b' * sy 'a' * sy 'c' == sy 'a' * sy 'b' ^ 2 * sy 'a' * sy 'c'
True

but they are not Commutative

>>> sy 'a' * sy 'b' ^ 2 * sy 'a' * sy 'c' == sy 'a' ^ 2 * sy 'b' ^ 2 * sy 'c'
False

and they admit a total right operation <* of Permutation N

>>> (sy 'a' * sy 'b' ^ 2 * sy 'a' * sy 'c') <* (pmtSwap 1 3 :: Permutation N)
ProductSymbol['a'^2*'b'^2*'c']

Note

Free products of symbols are finite complete sequences and allow a compact representation for repetitions and serve merely as dimensions for matrices (see OAlg.Entity.Matrix.Dim).

Possibly infinite complete sequences are represented by [x].

Constructors

ProductSymbol (Product N (U x))

sy :: Entity x => x -> ProductSymbol x

symbol of an entity, i.e. the complete sequence of psyLength one consisting just of it.

Example

>>> sy 'a'
ProductSymbol['a']

>>> sy 'a' * sy 'b' * sy 'b' ^ 5 * sy 'c'
ProductSymbol['a'*'b'^6*'c']

psyShow :: Entity x => ProductSymbol x -> String

showing as a product of symbols.

psyxs :: ProductSymbol x -> [(x, N)]

the indexed listing of the symbols.

psywrd :: Entity x => ProductSymbol x -> Word N x

the underlying word.

wrdpsy :: Entity x => Word N x -> ProductSymbol x

from word.

nProxy :: Proxy N

proxy for N.

psyJoin :: Entity x => ProductSymbol (ProductSymbol x) -> ProductSymbol x

joining complete sequences.

productSymbol :: Entity x => [x] -> ProductSymbol x

the induced product of symbols.

psyLength :: ProductSymbol x -> N

the length of a complete sequence.

psyFactor :: ProductSymbol x -> N -> x

the symbol for the given index.

psyMap :: Entity y => (x -> y) -> ProductSymbol x -> ProductSymbol y

mapping free products of symbols.

U

newtype U x

adjoins the point () to an entity.

Note Serves to build sums or products over symbols in x.

Constructors

U x

fromU :: U x -> x

deconstructor.

X

xProductSymbol :: Entity x => N -> X x -> X (ProductSymbol x)

random variable of complete sequences with the given maximal length.