OAlg.Entity.Sequence.Graph

Description

graphs of entities in x.

Graph

newtype Graph i x

mapping from an ordered index type i to a Entity type x.

Property Let g = Graph ixs be in Graph i x for a ordered Entity type i and Entity type x, then holds:

For all ..(i,_):(j,_).. in ixs holds: i < j.

lengthN g == lengthN ixs.

Constructors

Graph [(i, x)]

gphLength :: Graph i x -> N

the length of a graph.

gphxs :: Graph i x -> [(i, x)]

the underlying associations.

gphSqc :: (i -> Maybe x) -> Set i -> Graph i x

the induced graph given by a set of indices and a mapping.

gphLookup :: Ord i => Graph i x -> i -> Maybe x

looks up the mapping of an index.