OAlg.Entity.Sequence.PSequence

Description

partially defined sequences of items in x with a totally ordered index type i.

Sequence

newtype PSequence i x

partially defined sequences (x0,i0),(x1,i1).. of index items in x with a totally ordered index type i.

Property Let PSequence xis be in PSequence i x then holds: i < j for all ..(_,i):(_,j).. in xis.

Examples

>>> PSequence [('a',3),('b',7),('c',12)] :: PSequence N Char
PSequence [('a',3),('b',7),('c',12)]

and

>>> validate (valid (PSequence [('a',3),('b',7),('c',12)] :: PSequence N Char))
Valid

but

>>> validate (valid (PSequence [('a',3),('b',15),('c',12)] :: PSequence N Char))
Invalid

as Char is a totally ordered type it can serve as index type

>>> validate (valid (PSequence [(12,'c'),(3,'e'),(8,'x')] :: PSequence Char Z))
Valid

and they admit a total right operation <* of Permutation i

>>> (PSequence [(12,'c'),(3,'e'),(8,'x')] :: PSequence Char Z) <* pmtSwap 'e' 'x'
PSequence [(12,'c'),(8,'e'),(3,'x')]

Note As we keep the constructor public, it is crucial for there further use to ensure that they are valid!

Constructors

PSequence [(x, i)]

iProxy :: s i x -> Proxy i

proxy of the second type valiable i.

psqSpan :: Ord i => PSequence i x -> Span i

the span.

psqEmpty :: PSequence i x

the empty partially defined sequence.

psqIsEmpty :: PSequence i x -> Bool

checks of being empty.

psqxs :: PSequence i x -> [(x, i)]

the underlying list of indexed values.

psequence :: Ord i => (x -> x -> x) -> [(x, i)] -> PSequence i x

the partial sequenc given by a aggregation function an a list of value index pairs, which will be sorted and accordingly aggregated by thegiven aggregation function.

psqHead :: PSequence i x -> (x, i)

the head of a partial sequence.

psqTail :: PSequence i x -> PSequence i x

the tail.

psqMap :: (x -> y) -> PSequence i x -> PSequence i y

maps the entries, where the indices are preserved.

psqMapShift :: Number i => i -> ((x, i) -> y) -> PSequence i x -> PSequence i y

maps and shifts a partial sequence.

psqFilter :: (x -> Bool) -> PSequence i x -> PSequence i x

filters the partially defiend sequence accordingly the given predicate.

psqSplitWhile :: ((x, i) -> Bool) -> PSequence i x -> (PSequence i x, PSequence i x)

splits the sequence as long as the given predicate holds.

psqInterlace :: Ord i => (x -> y -> z) -> (x -> z) -> (y -> z) -> PSequence i x -> PSequence i y -> PSequence i z

interlaces the tow partially defined sequences according to the given mappings.

psqCompose :: (Ord i, Ord j) => PSequence i x -> PSequence j i -> PSequence j x

composition of the two partially defined sequences.

Property Let f be in PSequence i x and g be in PSequence j i then f `psqCompose` g is given by join . fmap ((??) f) . (??) g.

psqAppend :: PSequence i x -> PSequence i x -> PSequence i x

appends the second partially defined sequence to the first.

Property Let zs = psqAppend xs ys where ..(x,l) = xs and (y,f).. = ys then holds:

If

l < f

Then

zs is valid.

psqShear :: Ord i => (Maybe a -> Maybe a -> Maybe a, i) -> (Maybe a -> Maybe a -> Maybe a, i) -> PSequence i a -> PSequence i a

shears the two entries at the given position and leafs the others untouched.

Property Let x' = psqShear (sk,k) (sl,l) x, then holds

If

k < l

Then

x' k == sk (x k) (x l) and x' l == sl (x k) (x l).

x' i == x i for all i /= k, l.

psqSwap :: Ord i => i -> i -> PSequence i a -> PSequence i a

swaps the the k-th and the l-th entry.

Property Let x' = psqSwap k l x, then holds:

If

k < l

Then

x' k == x l and x' l == x k.

x' i == x i for all i /= k, l.

X

xPSequence :: Ord i => N -> N -> X x -> X i -> X (PSequence i x)

xPSequence n m random variable of partially defined sequences with maximal length min n m.