OAlg.Entity.Sequence.Set

Description

sets of ordered entities.

Set

newtype Set x

set of ordered entities in x.

Property Let s = Set xs be in Set x for a ordered Entity type x, then holds:

For all ..x:y.. in xs holds: x < y.

lengthN s == lengthN xs.

Constructors

Set [x]

set :: Ord x => [x] -> Set x

makes a set from the given list.

setSpan :: Set x -> Span x

the span of a set.

setxs :: Set x -> [x]

the elements of a set.

setSqc :: Ord x => (i -> Maybe x) -> Set i -> Set x

mapping a set.

setMap :: Ord y => (x -> y) -> Set x -> Set y

mapping of sets.

Note This works only for finite sets!

isSubSet :: Ord x => Set x -> Set x -> Bool

checks for being a sub set.

Operations

setEmpty :: Set x

the empty set.

setUnion :: Ord x => Set x -> Set x -> Set x

the union of two sets.

Lookup

setIndex :: Ord x => Set x -> x -> Maybe N

the index of an element, where the elements of the given set are indexed from 0.

Examples

>>> setIndex (Set ['a'..'x']) 'c'
Just 2

X

xSet :: Ord x => N -> X x -> X (Set x)

random variable of sets with maximal the given length.

Propositions

prpSetUnion :: (Ord x, Show x) => X (Set x) -> Statement

validity for the union operator of sets.