OAlg.Entity.Slice.Adjunction

Description

Cokernel-Kernel Adjunction for Sliced structures.

Adjunction

slcAdjunction :: (SliceCokernelTo i c, SliceKernelFrom i c) => i c -> Adjunction (SliceCokernelKernel i c) (SliceFactor From i c) (SliceFactor To i c)

the cokernel-kenrel adjunction.

Homomorphism

data SliceCokernelKernel i c x y where

the left and right homomorphisms for the cokernel-kernel adjunction slcAdjunction.

Constructors

SliceCokernel :: SliceCokernelKernel i c (SliceFactor To i c) (SliceFactor From i c)

SliceKernel :: SliceCokernelKernel i c (SliceFactor From i c) (SliceFactor To i c)

class (Distributive c, Sliced i c) => SliceCokernelTo i c where

Distributive structures c having to each Slice To i c a Cokernel.

Property Let h = SliceTo _ h' be in Slice To i c for a i sliced, Distributive structure c, then holds:

diagram (universalCone coker) == cokernelDiagram h' where coker = sliceCokernelTo h.

Methods

sliceCokernelTo :: Slice To i c -> Cokernel N1 c

class (Distributive c, Sliced i c) => SliceKernelFrom i c where

Distributive structures c having to each Slice From i c a Kernel.

Property Let h = SliceFrom _ h' be in Slice From i c for a i sliced, Distributive structure c, then holds:

diagram (universalCone ker) == kernelDiagram h' where coker = sliceKernelFrom h.

Methods

sliceKernelFrom :: Slice From i c -> Kernel N1 c

Unit

slcCokerKer :: (SliceCokernelTo i c, SliceKernelFrom i c) => Slice To i c -> SliceFactor To i c

the right unit of the cokernel-kernel adjunction slcAdjunction.

slcKerCoker :: (SliceCokernelTo i c, SliceKernelFrom i c) => Slice From i c -> SliceFactor From i c

the left unit of the cokernel-kenrel adjunction slcAdjunction.

X

xSliceFactorTo :: (Multiplicative c, Sliced i c) => XOrtSite To c -> i c -> X (SliceFactor To i c)

random variable for SliceFactor To i c.

xSliceFactorFrom :: (Multiplicative c, Sliced i c) => XOrtSite From c -> i c -> X (SliceFactor From i c)

random variable for SliceFactor From i c.

Proposition

prpHomMltSliceCokernelKernel :: (SliceCokernelTo i c, SliceKernelFrom i c) => XOrtSite To c -> XOrtSite From c -> i c -> Statement

validity for the values of SliceCokernelKernel to be HomMultiplicative.