OAlg.Entity.Slice.Definition

Description

definition of slicing a Multiplicative structures according a given indexed Point.

Slice

data Slice s i c where

slice over c by a given Site and indexed by i.

Constructors

SliceFrom :: i c -> c -> Slice From i c

SliceTo :: i c -> c -> Slice To i c

slice :: Slice s i c -> c

the underlying slice.

Factor

data SliceFactor s i c

factor between two slices.

Property Let SliceFactor a b f be in SliceFactor s i c for a Multiplicative structure c constrained by Sliced i c then holds:

If a matches SliceFrom _ a' then holds: Let SliceFrom _ b' = b in

orientation f == end a' :> end b'.

b' == f * a'.

If a matches SliceTo _ a' then holds: Let SliceTo _ b' = b in

orientation f == start a' :> start b'.

a' == b' * f .

Constructors

SliceFactor (Slice s i c) (Slice s i c) c

slfDrop :: SliceFactor s i c -> c

the underlying factor.

Sliced

class (Entity1 i, Singleton1 i) => Sliced i c where

Slicing a Multiplicative structures at the Point given by the type of the index i.

Note The constraint Singleton1 i ensures that the distinguished point depends only on the type i c.

Methods

slicePoint :: i c -> Point c

the distingueished point of the given index type i.

Hom

data SliceFactorDrop s x y where

dropping a slice factor.

Constructors

SliceFactorFromDrop :: (Multiplicative c, Sliced i c) => SliceFactorDrop From (SliceFactor From i c) c

SliceFactorToDrop :: (Multiplicative c, Sliced i c) => SliceFactorDrop To (SliceFactor To i c) c

Limes

data DiagramSlicedCenter i t n m c where

predicate for a Star t diagram with center Point given by the index type i c.

Property Let DiagramSlicedCenter i d be in DiagramSlicedCenter i t n m c then holds: slicePoint i == dgCenter d.

Constructors

DiagramSlicedCenter :: Sliced i c => i c -> Diagram (Star t) n m c -> DiagramSlicedCenter i t n m c

data LimesSlicedTip i s p t n m c where

predicate for a limes with a sliced tip of the universal cone.

Property Let LimesSlicedTip i l be in LimesSlicedTip i s p t n m c then holds: tip (universalCone l) == slicePoint i.

Constructors

LimesSlicedTip :: Sliced i c => i c -> Limes s p t n m c -> LimesSlicedTip i s p t n m c

lstLimes :: LimesSlicedTip i s p t n m c -> Limes s p t n m c

the underlying limes.

Projective

slfTerminalPoint :: (Multiplicative c, Sliced i c) => TerminalPoint (SliceFactor To i c)

terminal point for factors sliced to a Point.

slfPullback :: Multiplicative c => Products n (SliceFactor To i c) -> DiagramSlicedCenter i To (n + 1) n c -> Pullback n c

the induced pullback.

Injective

slfLimitsInjective :: (Multiplicative c, Sliced i c) => Limits Mlt Injective t n m c -> Limits Mlt Injective t n m (SliceFactor To i c)

the induced Injective Limits.

X

xSliceTo :: Sliced i c => XOrtSite To c -> i c -> X (Slice To i c)

the induced random variable.

xSliceFrom :: Sliced i c => XOrtSite From c -> i c -> X (Slice From i c)

the induced random variable.

xosXOrtSiteToSliceFactorTo :: (Multiplicative c, Sliced i c) => XOrtSite To c -> i c -> XOrtSite To (SliceFactor To i c)

the induced random variable.

xosXOrtSiteFromSliceFactorFrom :: (Multiplicative c, Sliced i c) => XOrtSite From c -> i c -> XOrtSite From (SliceFactor From i c)

the induced random variable.