OAlg.Hom.Additive

Description

homomorphisms between Additive structures

Additive

class (EmbeddableMorphism h Add, HomFibred h) => HomAdditive h

type family of homomorphisms between Additive structures.

Property Let h be a type instance of the class HomAdditive, then for all a, b and f in h a b holds:

For all r in Root a holds: amap f (zero r) == zero (rmap f r).

For all x and y in a with root x == root y holds: amap f (x + y) ) == amap f x + amap f y.

Such a h will be called a family of homomorphisms between additive structures and an entity f of h a b a additive homomorphism between a and b.

Proposition

prpHomAdd1 :: HomAdditive h => h a b -> Root a -> Statement

validity according to OAlg.Hom.Additive.

prpHomAdd2 :: HomAdditive h => h a b -> Adbl2 a -> Statement

validity according to OAlg.Hom.Additive.