OAlg.Hom.Algebraic

Description

homomorphisms between Algebraic structures having the same associated Scalar.

Algebraic

class (EmbeddableMorphism h (Alg k), HomDistributive h, HomVectorial k h) => HomAlgebraic k h

type family of homomorphisms between Algebraic structures having the same associated Scalar.