OAlg.Hom.Definition

Description

introducing the idiom Hom.

Hom

type family Hom s (h :: Type -> Type -> Type) :: Constraint

parameterized constraint that the values of the type h x y admit the constraints of a homomorphisms between the structures given by s.