OAlg.Hom.Fibred

Description

homomorphisms between Fibred structures

Fibred

class (EmbeddableMorphism h Fbr, Applicative h, Entity2 h, EmbeddableMorphismTyp h) => HomFibred h where

type family of homomorphisms between Fibred structures.

Property Let h be an instance of HomFibred then for all a, b and f in h a b and x in a holds: root (amap f x) == rmap f (root x).

Minimal complete definition

Nothing

Methods

rmap :: h a b -> Root a -> Root b

default rmap :: (EmbeddableMorphism h FbrOrt, HomOriented h) => h a b -> Root a -> Root b

class (Category h, Functorial h, HomFibred h) => FunctorialHomFibred h

functorial application of Fibred homomorphisms.

Fibred Oriented

class (EmbeddableMorphism h FbrOrt, HomFibred h, HomOriented h) => HomFibredOriented h

type family of homomorphisms between FibredOriented structures.

Property Let h be an instance of HomFibredOriented then for all a, b and f in h a b and r in Root a holds: rmap f r == omap f r.

Proposition

prpHomFbrOrt :: HomFibredOriented h => h a b -> Root a -> Statement

validity according to HomFibredOriented.