OAlg.Hom.Oriented.Definition

Description

definition of homomorphisms between Oriented structures.

Homomorphism

class (EmbeddableMorphism h Ort, Applicative h, Entity2 h, EmbeddableMorphismTyp h, Transformable1 Op (ObjectClass h)) => HomOriented h where

type family of homomorphisms between Oriented structures.

Property Let h be an instance of HomOriented, then for all a, b and f in h a b and x in a holds: start (amap f x) == pmap f (start x) and end (amap f x) == pmap f (end x).

We call such a h a family of homomorphisms between oriented structures and an entity f in h a b a covariant oriented homomorphism - or oriented homomorphism for short - between a and b. A covariant oriented homomorphism f in h (Op a) b or h a (Op b) will be called a contravariant oriented homomorphism between a and b.

Note

As h is an instance of EmbeddableMorphism h Ort it follows that for all a, b and f in h a b holds: tauHom (homomorphous f) :: Homomorphous Ort a b and thus a and b are Oriented structures! How to work with this concretely see the implementation of prpHomOrt where the property above is stated.

The constraint EmbeddableMorphismTyp for a family h of homomorphisms between Oriented structures ensures that the type Path h is a instances of Eq2.

Methods

pmap :: h a b -> Point a -> Point b

omap :: HomOriented h => h a b -> Orientation (Point a) -> Orientation (Point b)

the induced mapping of Orientation.

type IsoOrt s h = (FunctorialHomOriented h, Cayleyan2 h, Hom s h)

s-isomoprhisms.

type IsoOriented h = (FunctorialHomOriented h, Cayleyan2 h)

isomorphisms between Oriented structures.

Functorial

class (Category h, Functorial h, HomOriented h) => FunctorialHomOriented h

functorial application on Oriented structures.

Properties Let h be an instance of the class FunctorialHomOriented, then holds:

For all types a and s in Struct (ObjectClass h) a holds: pmap (cOne s) = id.

For all types a, b, c and f in h b c, g in h a b holds: pmap (f . g) = pmap f . pmap g.

IdHom

data IdHom s a b where

identity morphism.

Constructors

IdHom :: Structure s a => IdHom s a a

OpHom

data OpHom h x y where

induced homomorphism on Op.

Constructors

OpHom :: Transformable1 Op (ObjectClass h) => h x y -> OpHom h (Op x) (Op y)

HomOp

data HomOp s a b where

some basic contravariant isomorphisms between s-structures with there invert2.

Constructors

FromOpOp :: (Structure s (Op (Op a)), Structure s a) => HomOp s (Op (Op a)) a

ToOpOp :: (Structure s (Op (Op a)), Structure s a) => HomOp s a (Op (Op a))

OpPath :: (Structure s a, Structure s (Op (Path a)), Structure s (Path (Op a))) => HomOp s (Op (Path a)) (Path (Op a))

OpPathInv :: (Structure s a, Structure s (Op (Path a)), Structure s (Path (Op a))) => HomOp s (Path (Op a)) (Op (Path a))

Opposite :: (Structure s (Op (Orientation p)), Structure s (Orientation p)) => HomOp s (Op (Orientation p)) (Orientation p)

OppositeInv :: (Structure s (Op (Orientation p)), Structure s (Orientation p)) => HomOp s (Orientation p) (Op (Orientation p))

IsoOp

data IsoOp s a b

isomorphisms induced by paths of HomOp.

type PathHomOp s a b = Path (HomOp s) a b

paths of HomOp.

opPathOrt :: Oriented a => IsoOp Ort (Op (Path a)) (Path (Op a))

the induced isomorphism given by OpPath.

isoFromOpOpOrt :: Oriented a => IsoOp Ort (Op (Op a)) a

the induced isomorphism of Oriented structures given by FromOpOp.

Examples

let tOS = invert2 (isoFromOpOpOrt :: IsoOp Ort (Op (Op OS)) OS)
let f = isoFromOpOpOrt :: Oriented a =>IsoOp Ort (Op (Op a)) a
let t = invert2 f

>>> tOS
IsoOp Path[ToOpOp]

>>> t . t . tOS
IsoOp Path[ToOpOp,ToOpOp,ToOpOp]

>>> f . f . t . f . t . tOS
IsoOp Path[]

>>> f . f . t . f . t . tOS == cOne Struct
True

IsoOpMap

data IsoOpMap f s a b

isomorphisms induced by paths of OpMap.

type PathOpMap f s = Path (OpMap f s)

paths of OpMap.

data OpMap f s a b where

contravariant s-isomorphisms between f x and f (Op x).

Constructors

ToOp1 :: (Structure s (Op (f x)), Structure s (f (Op x)), Structure s x) => OpMap f s (Op (f x)) (f (Op x))
contravariant s-isomorphism from f x to f (Op x).

FromOp1 :: (Structure s (Op (f x)), Structure s (f (Op x)), Structure s x) => OpMap f s (f (Op x)) (Op (f x))
the inverse of ToOp1.

toOp1Struct :: OpMap f s (Op (f x)) (f (Op x)) -> Struct s x

structural attest for ToOp1.

fromOp1Struct :: OpMap f s (f (Op x)) (Op (f x)) -> Struct s x

structural attest for FromOp1.

Path

isoCoPath :: Oriented x => IsoOpMap Path Ort (Op (Path x)) (Path (Op x))

contravariant isomorphism from Path x to Path (Opx) .