OAlg.Hom.Vectorial

Description

homomorphisms between Vectorial structures having the same associated Scalar.

Vectorial

class (EmbeddableMorphism h (Vec k), HomAdditive h) => HomVectorial k h

type family of homomorphisms between Vectorial structures having the same associated 'Scalar.

Property Let h be a type instance of the class HomVectorial k, then for all a, b, v in h a b and x in k holds: amap h (x ! v) == x ! amap h v.