OAlg.Limes.KernelsAndCokernels

Description

kernels and cokernels, i.e. limits in a Distributive structure of Diagram (Parallel d) making all arrows zero.

Kernels

type Kernels n = Limits Dst Projective (Parallel LeftToRight) N2 n

kernels for Distributive structures.

type Kernel n = Limes Dst Projective (Parallel LeftToRight) N2 n

kernel as a Limes.

type KernelCone n = Cone Dst Projective (Parallel LeftToRight) N2 n

Cone for a kernel.

type KernelDiagram n = Diagram (Parallel LeftToRight) N2 n

Diagram for a kernel.

kernelFactor :: KernelCone N1 c -> c

the factor of its shell.

kernelDiagram :: Oriented c => c -> KernelDiagram N1 c

the kernel diagram of a given factor.

Construction

kernels :: Distributive a => Kernels N1 a -> Kernels n a

promoting kernels.

kernels0 :: Distributive a => Kernels N0 a

kernels for zero arrows.

kernels1 :: Distributive a => Kernels N1 a -> Kernels (n + 1) a

promoting kernels.

krnEqls :: (Distributive a, Abelian a) => Kernels n a -> Equalizers (n + 1) a

the induced equalizers where its first arrow is zero.

eqlKrns :: Distributive a => Equalizers (n + 1) a -> Kernels n a

the induced kernels given by adjoining a zero arrow as first arrow.

kernelZero :: Distributive c => p c -> Orientation (Point c) -> Kernel N1 c

the kernel of the zero factor given by the orientation, i.e. one

Orientation

kernelsOrnt :: Entity p => p -> Kernels n (Orientation p)

kernels for Orientation.

Cokernels

type Cokernels n = Limits Dst Injective (Parallel RightToLeft) N2 n

cokernels for Distributive structures.

type Cokernel n = Limes Dst Injective (Parallel RightToLeft) N2 n

cokernel as Limes.

type CokernelCone n = Cone Dst Injective (Parallel RightToLeft) N2 n

Cone for a cokernel.

type CokernelDiagram n = Diagram (Parallel RightToLeft) N2 n

Diagram for a cokernel.

cokernelFactor :: CokernelCone N1 c -> c

the factor of its shell.

cokernelDiagram :: Oriented c => c -> CokernelDiagram N1 c

the cokernel diagram of a given factor.

Construction

cokernels :: Distributive a => Cokernels N1 a -> Cokernels n a

promoting cokernels.

cokernels' :: Distributive a => p n -> Cokernels N1 a -> Cokernels n a

cokernels given by an additional proxy for n.

Orientation

cokernelsOrnt :: Entity p => p -> Cokernels n (Orientation p)

cokernels for Orientation.

Duality

krnLimesDuality :: Distributive a => LimesDuality Dst (Kernel n) (Cokernel n) a

duality from Kernel to Cokernel.

cokrnLimesDuality :: Distributive a => LimesDuality Dst (Cokernel n) (Kernel n) a

duality between Cokernel to Kernel.

Proposition

prpIsKernel :: Distributive a => Kernel n a -> FinList n a -> a -> Statement

checks if the arrows of the kernel diagram are equal to the given ones and if its shell is equal to the given arrow.

Property Let LimesProjective ('ConeKerenl d k') _ = ker be in Kernel n a, fs in FinList n a and k be in a, then the statement prpIsKernel ker fs k holds if and only if

fs == dgArrows d.

k == k'.

prpIsCokernel :: Distributive a => Cokernel n a -> FinList n a -> a -> Statement

checks if the arrows of the cokernel diagram are equal to the given ones and if its shell is equal to the given arrow.

Property Let LimesInjective ('ConeCokerenl d k') _ = coker be in Cokernel n a, fs in FinList n a and k be in a, then the statement prpIsCokernel coker fs k holds if and only if

fs == dgArrows d.

k == k'.