OAlg.Limes.MinimaAndMaxima

Description

minima and maxima within a Multiplicative structure, i.e. limits of Diagram (Chain t).

Minima

type Minima t n = Limits Mlt Projective (Chain t) (n + 1) n

minima for a Multiplicative structure.

type Minimum t n = Limes Mlt Projective (Chain t) (n + 1) n

minimum as Limes.

type MinimumCone t n = Cone Mlt Projective (Chain t) (n + 1) n

Cone for a minimum.

type MinimumDiagram t n = Diagram (Chain t) (n + 1) n

Diagram for a minimum.

minimaTo :: Multiplicative a => Minima To n a

minima according to Chain To.

minimaFrom :: Multiplicative a => Minima From n a

minima according to Chain From.

Maxima

type Maxima t n = Limits Mlt Injective (Chain t) (n + 1) n

maxima for a Multiplicative structure.

type Maximum t n = Limes Mlt Injective (Chain t) (n + 1) n

maximum as a Limes.

type MaximumCone t n = Cone Mlt Injective (Chain t) (n + 1) n

Cone for a maximum.

type MaximumDiagram t n = Diagram (Chain t) (n + 1) n

Diagram for a maximum.

maximaTo :: Multiplicative a => Maxima To n a

maxima according to Chain To.

maximaTo' :: Multiplicative a => p n -> f a -> Maxima To n a

maxima according to Chain To given by two proxy types.

maximaFrom :: Multiplicative a => Maxima From n a

maxima according to Chain From.

maximaFrom' :: Multiplicative a => p n -> f a -> Maxima From n a

maxima according to Chain From given by two proxy types.

Duality

maxLimitsDualityTo :: Multiplicative a => LimitsDuality Mlt (Maxima To n) (Minima From n) a

duality between Maxima To and Minima From.

maxLimitsDualityFrom :: Multiplicative a => LimitsDuality Mlt (Maxima From n) (Minima To n) a

duality between Maxima From and Minima To.