OAlg.Limes.ProductsAndSums

Description

products and sums, i.e. limits of Diagram Discrete.

Products

type Products n = Limits Mlt Projective Discrete n N0

products for a Multiplicative structure.

type Product n = Limes Mlt Projective Discrete n N0

product as a Limes.

type ProductCone n = Cone Mlt Projective Discrete n N0

Cone for a product.

type ProductDiagram n = Diagram Discrete n N0

Diagram for a product.

Construction

prdDiagram :: Oriented a => Diagram (Star From) (n + 1) n a -> ProductDiagram n a

the underlying product diagram.

prdCone :: Multiplicative a => Diagram (Star From) (n + 1) n a -> ProductCone n a

the product cone.

products :: Multiplicative a => Products N0 a -> Products N2 a -> Products n a

products of n points given by products of zero and two points.

products0 :: Multiplicative a => TerminalPoint a -> Products N0 a

products of zero points given by a terminal point.

products1 :: Multiplicative a => Products N1 a

products of one point, i.e. Minima.

products2 :: Multiplicative a => Products N2 a -> Products (n + 2) a

products of at least two points given by products of two points.

Orientation

prdConeOrnt :: Entity p => p -> FinList n p -> ProductCone n (Orientation p)

product cone for Orientation.

productOrnt :: Entity p => p -> FinList n p -> Product n (Orientation p)

product for Orientation.

productsOrnt :: Entity p => p -> Products n (Orientation p)

products for Orientation.

Sums

type Sums n = Limits Mlt Injective Discrete n N0

sums for a Multiplicative structure.

type Sum n = Limes Mlt Injective Discrete n N0

sum as a 'Limes.

type SumCone n = Cone Mlt Injective Discrete n N0

Cone for a sum.

type SumDiagram n = Diagram Discrete n N0

Diagram for a sum.

Duality

sumLimitsDuality :: Multiplicative a => LimitsDuality Mlt (Sums n) (Products n) a

duality between sums and products.

Construction

sumDiagram :: Oriented a => Diagram (Star To) (n + 1) n a -> SumDiagram n a

the underlying sum diagram given by a sink diagram.

sumCone :: Multiplicative a => Diagram (Star To) (n + 1) n a -> SumCone n a

the sum cone given by a sink diagram.

sums :: Multiplicative a => Sums N0 a -> Sums N2 a -> Sums n a

sums of n points given by sums of zero and two points.

sums' :: Multiplicative a => p n -> Sums N0 a -> Sums N2 a -> Sums n a

sums given by a proxy type for n.

sums0 :: Multiplicative a => InitialPoint a -> Sums N0 a

sums of zero points given by a initial point.

sums1 :: Multiplicative a => Sums N1 a

sums of one point, i.e. Maxima.

sums2 :: Multiplicative a => Sums N2 a -> Sums (n + 2) a

sums of at least two points given by sums of two points.

Orientation

sumConeOrnt :: Entity p => p -> FinList n p -> SumCone n (Orientation p)

sum cone for Orientation.

sumOrnt :: Entity p => p -> FinList n p -> Sum n (Orientation p)

sum for Orientation.

sumsOrnt :: Entity p => p -> Sums n (Orientation p)

sums for Orientation.