OAlg.Structure.Additive.Definition

Description

additive structures, i.e. structures with a partially defined addition (+).

Additive

class Fibred a => Additive a where

Fibred structures with a partialy defined addition and having zero as the neutral element of the summation. An entity of a Additive structure will be called a summand.

Properties Let a be a Additive structure, then holds:

For all r in Root a holds: root (zero r) == r.

For all f and g in a holds:

If root f == root g then f + g is valid and root (f + g) == root f.

If root f /= root g then f + g is not valid and its evaluation will end up in a NotAddable-exception.

For all f, g in a with root f == root g holds: f + g == g + f.

For all f in a holds: f + zero (root f) == f

For all f, g, h in a with root f == root g == root h holds: (f + g) + h == f + (g + h).

For all f in a and n in N holds: ntimes 0 f == zero (root f) and ntimes (n + 1) f == f + ntimes n f.

Minimal complete definition

zero, (+)

Methods

zero :: Root a -> a

the neutral element associated to each root. If there is no ambiguity for zero r we will briefly denote it by 0 r or just 0.

(+) :: a -> a -> a infixl 6

the addition for two summands.

ntimes :: N -> a -> a

n times of a summand.

zero' :: Additive a => p a -> Root a -> a

the zero to a given root. The type p c serves only as proxy and zero' is lazy in it.

Note As Point may be a non-injective type family, the type checker needs some times a little bit more information to pic the right zero.

data Add

type representing the class of Additive structures.

class (ForgetfulFbr s, Transformable s Add) => ForgetfulAdd s

transformable to Additive structure.

Abelian

class Additive a => Abelian a where

Additive structures having for each summand an additve inverse.

Properties Let a be a Additive structure, then holds:

For all f in a holds: root (negate f) == root f.

For all f in a holds: f + negate f == zero (root f).

For all f and g in a holds:

If root f == root g then f - g is valid and root (f - g) == root f.

If root f /= root g then f - g is not valid and its evaluation will end up in a NotAddable-exception.

For f and g in a with root f == root g holds: f - g == f + negate g.

For all z in Z and f in a holds:

If 0 <= z then ztimes z f == ntimes (prj z) f.

If z < 0 then ztimes z f == negate (ntimes (prj z) f).

Minimal complete definition

negate | (-)

Methods

negate :: a -> a

negation of a summand.

(-) :: a -> a -> a infixl 6

subtraction of two summands.

Properties

ztimes :: Z -> a -> a

z times of a sumand.

isZero :: Additive a => a -> Bool

check for beeing zero.

data Abl

type representing the class of Abelian structures.

class (ForgetfulFbr s, ForgetfulAdd s, Transformable s Abl) => ForgetfulAbl s

transformable to Abelian structure.

Orphan instances

Additive a => Projectible a (Sheaf a)

Instance details

Methods

prj :: Sheaf a -> a