OAlg.Structure.Distributive.Proposition

Description

propositions about Distributive structure.

Distributive

prpDst :: Distributive d => XDst d -> Statement

validity for the Distributive structure of the type a.

data XDst d

random variable for validating Distributive structures.

Constructors

XDst (X (DstRootSide LeftSide d)) (X (DstSide LeftSide d)) (X (DstRootSide RightSide d)) (X (DstSide RightSide d))

prpDst1 :: Distributive d => X (DstRootSide LeftSide d) -> Statement

validity according to OAlg.Structure.Distributive.

prpDst2 :: Distributive d => X (DstSide LeftSide d) -> Statement

validity according to OAlg.Structure.Distributive.

prpDst3 :: Distributive d => X (DstRootSide RightSide d) -> Statement

validity according to OAlg.Structure.Distributive.

prpDst4 :: Distributive d => X (DstSide RightSide d) -> Statement

validity according to OAlg.Structure.Distributive.

data DstRootSide (s :: Side) d where

predicate for a root and a factor at the root.

Properties Let p be in DstRootSide s d for a Distributive structure d, then holds:

If p matches LDstRoot r f then holds: start r == end f.

If p matches RDstRoot f r then holds: start f == end r.

Constructors

LDstRoot :: Root d -> d -> DstRootSide LeftSide d

RDstRoot :: d -> Root d -> DstRootSide RightSide d

data DstSide (s :: Side) d where

predicate for two addable summands together with a matching factor.

Properties Let p be in DstSide s d for a Distributive structure, then holds:

If p matches LDst (a,b) g then holds:

root a == root b.

start a == end g.

If p matches RDst f (a,b) then holds:

root a == root b.

start f == end a.

Constructors

LDst :: (d, d) -> d -> DstSide LeftSide d

RDst :: d -> (d, d) -> DstSide RightSide d

X

class XStandardDst d where

standard random variable for Distributive structures.

Methods

xStandardDst :: XDst d

xDstStalkStartEnd :: Distributive m => XStalk m -> XOrtSite From m -> XOrtSite To m -> XDst m

the induced random variable for Distributive structures.

Total

xDstTtl :: Singleton (Root m) => X m -> XDst m

random variable to validate Total Distributive structures.

Orientation

xDstOrnt :: Entity p => X p -> XDst (Orientation p)

random variable for the Distributive structure of Orientation p.

Oriented Direction

xoDst :: Distributive d => XOrtOrientation d -> XOrtSite From d -> XOrtSite To d -> XDst d

the induced random variable for Distributive structures.

Tools

dstDst :: Distributive d => Int -> XDst d -> IO ()

distribution of XDst d.