OAlg.Structure.Oriented.Definition

Description

definition of Oriented structures.

Oriented

class (Entity q, Entity (Point q)) => Oriented q where

types with a Oriented structure. The values of an Oriented structure will be called arrows and the values of the associated Point type points. To each arrow there is a start and a end point assigned.

Property Let q be a type instance of the class Oriented, then holds:

For all a in q holds: orientation a == start a :> end a.

Note

If the types q and Point q are interpreted as sets A and P and start, end as functions from A to P then this structure forms a quiver with arrows in A and points in P.

Morphisms can be interpreted as Oriented structures via SomeMorphism. The bad thing about this is that we lose the check for composability of two Morphisms given by the type checker, but we gain all the functionality of Oriented structures, i.e we can define homomorphisms, limits etc on Morphisms.

Minimal complete definition

orientation | start, end

Associated Types

type Point q

the associated type of points.

Methods

orientation :: q -> Orientation (Point q)

the orientation of an arrow.

start :: q -> Point q

the start point of an arrow.

end :: q -> Point q

the end point of an arrow.

class Singleton (Point x) => Total x

structures where its associated Point type is singleton. They yield globally defiend algebraic operations.

class Entity (Point x) => EntityPoint x

helper class to avoid undecidable instances.

class Ord (Point x) => OrdPoint x

helper class to circumvent undecidable instances.

isEndo :: Oriented q => q -> Bool

check for being an endo.

Definition Let q be a Oriented structure, then an arrow a in q is called endo if and only if start a == end a.

isEndoAt :: Oriented a => Point a -> a -> Bool

check for being an endo at the given point.

type OS = Orientation Symbol

as Orientation p is an instance of almost every structured class it serves as a standard type for validating.

data Ort

type representing the class of Oriented structures.

structOrtOp :: Struct Ort x -> Struct Ort (Op x)

attest that if x is Oriented then also Op x is Oriented.

class Transformable s Ort => ForgetfulOrt s

transformable to Oriented structure.

Transposable

class (Transposable q, Oriented q) => TransposableOriented q

transposable oriented structures.

Property Let q be a TransposableOriented structure, then holds: For all a in q holds: orientation (transpose a) == opposite (orientation a).

Orientation

data Orientation p

orientation given by the start point as its first component and the end point as its second.

Property For all o in Orientation p holds: o == start o :> end o.

Note As Orientations are instances of almost all algebraic structures defined here, they serve as a proof that this structures are instanceable.

Constructors

p :> p infix 5

opposite :: Orientation p -> Orientation p

the opposite orientation.

Path

data Path q

a path in a Oriented structure q starting at a given point.

Definition Let q be a Oriented structure and p = Path s [a 0..a (n-1)] a path in q, then p is valid if and only if

s is valid and a i are valid for all i = 0..n-1.

start (a (n-1)) == s and start (a i) == end (a (n+1)) for all i = 0..n-2.

furthermore n is called the length of p.

Note Paths admit a canonical embedding in to Product.

Constructors

Path (Point q) [q]

pthLength :: Path q -> N

the length of a path.

pthOne :: Point q -> Path q

path of length 0 at the given point.

pthMlt :: Oriented q => Path q -> Path q -> Path q

composition of two paths.

X

Site

data XOrtSite s q where

random variables X q and X (Point q) for Oriented structure q.

Properties Let q be an instance of the class Oriented, then holds:

Let XStart xp xStart be in XOrtSite From q, then holds: For all p in Point q and x in the range of xStart p holds: start x == p.

Let XEnd xp xEnd be in XOrtSite To q, then holds: For all p in Point q and x in the range of xEnd p holds: end x == p.

Note The random variables xp should have a bias to non trivial random variables xp >>= xStart or xp >>= xEnd.

Constructors

XStart :: X (Point q) -> (Point q -> X q) -> XOrtSite From q

XEnd :: X (Point q) -> (Point q -> X q) -> XOrtSite To q

class XStandardOrtSite s a where

standard random variable for XOrtSite.

Methods

xStandardOrtSite :: XOrtSite s a

class XStandardOrtSite To a => XStandardOrtSiteTo a

standard random variable for XOrtSite To, helper class to avoid undecidable instances.

class XStandardOrtSite From a => XStandardOrtSiteFrom a

standard random variable for XOrtSite From, helper class to avoid undecidable instances.

coXOrtSite :: XOrtSite s q -> Dual (XOrtSite s q)

to the dual of a XOrtSite s q, with inverse coXOrtSiteInv.

coXOrtSiteInv :: (Dual (Dual s) :~: s) -> Dual (XOrtSite s q) -> XOrtSite s q

from the dual of a Dual (XOrtSite s q), with inverse coXOrtSite.

xosFromOpOp :: XOrtSite s (Op (Op q)) -> XOrtSite s q

from the bidual.

xosStart :: XOrtSite From q -> Point q -> X q

the random variable of arrows in q having all as start the given point.

xosEnd :: XOrtSite To q -> Point q -> X q

the random variable of arrows in q having all as end the given point.

xosPathMaxAt :: Oriented q => XOrtSite s q -> N -> Point q -> X (Path q)

tries to make a path at the given point with maximal length of the given length.

Properties Let xPath = xosPathMaxAt xos n x, then holds:

If xos matches XStart _ xq then for all p in the range of xPath holds:

start p == x.

If pthLength p < n then xq (end p) matches XEmpty.

If xos matches XEnd _ xq then for all p in the range of xPath holds:

end p == x.

If pthLength p < n then xq (start p) matches XEmpty.

xosPathMax :: Oriented q => XOrtSite s q -> N -> X (Path q)

random variable of paths with maximal length of the given length.

Orientation

data XOrtOrientation q

random variable of arrows given by an orientation.

Properties Let XOrtOrientation xo xArrow be in XOrtOrientation q for a Oriented structure q, then holds: For all o in Orientation q and x in the range of xArrow o holds: orientation x == o.

Note The random variable xo should have a bias to non trivial random variables xo >>= xArrow and as such the range of xo should be included in one connection component of q.

Constructors

XOrtOrientation (X (Orientation (Point q))) (Orientation (Point q) -> X q)

xoOrientation :: XOrtOrientation q -> X (Orientation (Point q))

the underlying random variable of orientations.

xoArrow :: XOrtOrientation q -> Orientation (Point q) -> X q

the underlying random variable of arrow given by the orientation.

xoPoint :: Oriented q => XOrtOrientation q -> X (Point q)

the underlying random variable of points, i.e. the union of the induced start and end random variable of xoOrientation.

coXOrtOrientation :: XOrtOrientation q -> Dual (XOrtOrientation q)

to the dual.

xoTo :: Oriented q => XOrtOrientation q -> XOrtSite To q

the induced XOrtSite To

xoFrom :: Oriented q => XOrtOrientation q -> XOrtSite From q

the induced XOrtSite From.

xoTtl :: Total q => X q -> XOrtOrientation q

random variable of XOrtOrientation q for a total q.

xoOrnt :: X p -> XOrtOrientation (Orientation p)

the induced random variable of Orientation q.

class XStandardOrtOrientation q where

standard random variable for XOrtOrientation.

Methods

xStandardOrtOrientation :: XOrtOrientation q

Orientation

class XStandard (Point a) => XStandardPoint a

standard random variable of Points of a.

xStartOrnt :: X p -> XOrtSite From (Orientation p)

the XOrtSite From for Orientation p of the given random variable.

xEndOrnt :: X p -> XOrtSite To (Orientation p)

the XOrtSite To of Orientation p of the given random variable.