OAlg.Structure.Oriented.Proposition

Description

Propositions on Oriented structure.

Oriented

prpOrt :: Oriented q => XOrt q -> Statement

validity of the Oriented structure of q.

type XOrt = X

random variable for Oriented structures.

prpOrt0 :: Oriented q => X q -> Statement

validity of the functions orientation, start and end.

prpOrt1 :: Oriented q => X q -> Statement

validity of the relation between orientation, start and end according to OAlg.Structure.Oriented.Definition.

X

xosOrt :: Oriented q => XOrtSite s q -> XOrt q

the underlying random variable for Oriented structures.

xosPoint :: XOrtSite s q -> X (Point q)

the underlying random variable of Point g.

xoOrt :: XOrtOrientation q -> XOrt q

the underlying random variable for Oriented structures.

Orientation

xOrtOrnt :: X p -> XOrt (Orientation p)

the induced random variable of Oriented structures for Orientation p.